Testing & Bug fixes
