+++ app/src/mixins/DexInteraction.js
@@ -38,11 +38,10 @@ export default {
       if (depositCoin) this.depositCoin = depositCoin
       if (destinationCoin) this.destinationCoin = destinationCoin
     },
-    getAllCoins () {
-      let coins = this.$store.state.settings.coins.coinswitch.concat(this.$store.state.settings.coins.oneinch).concat(this.$store.state.settings.coins.defibox)
-
+    getUniqueTokens (coins) {
       let duplicates = []
-      coins = coins.map((el) => {
+
+      return coins.map((el) => {
         let search = coins.filter(o => el.value.toLowerCase() === o.value.toLowerCase())
         let value = search[0]
 
@@ -52,19 +51,26 @@ export default {
           value = null
         }
         return value
-      }).filter(o => o != null && !this.$store.state.wallets.tokens.filter(x => x.chain === 'eos').map(w => w.type).includes(o.value.toLowerCase()))
+      }).filter(o => o != null)
+    },
+    getAllCoins () {
+      let coins = this.$store.state.settings.coins.coinswitch.concat(this.$store.state.settings.coins.oneinch).concat(this.$store.state.settings.coins.defibox)
 
+      coins = this.getUniqueTokens(coins).filter(o => !(this.$store.state.wallets.tokens.filter(x => x.chain === 'eos').map(w => w.type.toLowerCase()).includes(o.value.toLowerCase())))
+      console.log(this.$store.state.wallets.tokens)
       this.$store.state.wallets.tokens.filter(o => o.chain === 'eos').forEach((coin) => {
         let row = {
           'label': coin.type,
           'name': coin.name,
           'value': coin.type,
+          'contract': coin.contract,
+          'precision': coin.precision,
           'image': 'https://ndi.340wan.com/eos/' + coin.contract + '-' + coin.type + '.png',
           'dex': 'coinswitch',
-          'amount': coin.amount,
+          'amount': parseFloat(coin.amount),
           'amountUSD': coin.usd
         }
-        console.log(coin)
+
         coins.unshift(row)
       })
 
@@ -87,7 +93,7 @@ export default {
 
       console.log(this.dex, ' this.dex')
       if (!this.dex) {
-        this.error = 'Cannot swap ' + this.depositCoin.value + ' to ' + this.destinationCoin.value
+        this.error = 'Cannot swap ' + this.depositCoin.value.toUpperCase() + ' to ' + this.destinationCoin.value
       } else {
         this.$store.commit('settings/setDex', { dex: this.dex, destinationCoin: this.destinationCoin, depositCoin: this.depositCoin })
       }
