--- app/src/pages/Exchange/Coin_switch.vue
@@ -1,2885 +1,0 @@
-<template>
-<div class="column text-black bg-grey-12" :class="screenSize > 1024 ? 'desktop-marg': 'mobile-pad'">
-    <!-- padding-bottom: 100px;background: #f3f3f3 !important -->
-    <div v-if="screenSize > 1024">
-        <div class="row">
-
-            <div class="col col-md-12">
-                <div class="desktop-card-style apps-section q-mb-sm">
-
-                    <!-- 1inch component -->
-                    <!-- add your code here -->
-
-                    <!-- Vdex component -->
-                    <div class="chain-tools-wrapper">
-                        <div class="standard-content"></div>
-                        <div class="chain-tools-wrapper--list open">
-                            <div class="list-wrapper">
-                                <div class="row">
-                                    <div class="col col-12">
-                                        <div class="trade-component">
-                                            <!-- <img src="statics/theme1/Screenshot_208.png" alt="" style="opacity: .1"> -->
-                                            <div v-if="pStep === 1" class="prototype">
-                                                <div class="head">Market Order</div>
-                                                <div class="you-pay">
-                                                    <div class="you-pay-head row items-center">
-                                                        <div class="col col-6">You Pay</div>
-                                                        <!-- <div class="col col-6 red text-right text-red">Max 0 USDT</div> -->
-                                                    </div>
-                                                    <div class="you-pay-body row items-center">
-                                                        <div class="col col-12 pay-coin">
-                                                            <!-- class="select-input" -->
-                                                            <span class="paycoin-search hidden">Type BTC</span>
-                                                            <span class="receivecoin-search hidden">Type BTC</span>
-                                                            <q-select filled options-cover stack-label class="pay-coin-select" popup-content-class="pay-coin-select-popup" v-model="depositCoin" @filter="filterDepositCoin" @input="checkGetPairs()" :disabled="!depositCoinOptions" :loading="!depositCoinOptions" :options="depositCoinOptions">
-                                                                <template v-slot:option="scope">
-                                                                    <q-item class="custom-menu" v-bind="scope.itemProps" v-on="scope.itemEvents">
-                                                                        <q-item-section avatar>
-                                                                            <q-icon class="option--avatar option--avatar__custom" :name="`img:${scope.opt.image}`" />
-                                                                        </q-item-section>
-                                                                        <q-item-section dark>
-                                                                            <q-item-label caption>{{ scope.opt.value }}</q-item-label>
-                                                                            <q-item-label v-html="scope.opt.label" />
-                                                                        </q-item-section>
-                                                                    </q-item>
-                                                                </template>
-                                                                <template v-slot:selected>
-                                                                    <q-item v-if="depositCoin">
-                                                                        <q-item-section avatar>
-                                                                            <q-icon class="option--avatar option--avatar__custom" :name="`img:${depositCoin.image}`" />
-                                                                        </q-item-section>
-                                                                        <q-item-section>
-                                                                            <q-item-label caption>{{ depositCoin.value }}</q-item-label>
-                                                                            <q-item-label v-html="depositCoin.label" />
-                                                                        </q-item-section>
-                                                                    </q-item>
-                                                                    <q-item v-else>
-                                                                    </q-item>
-                                                                </template>
-                                                            </q-select>
-                                                            <q-select filled options-cover stack-label class="receive-coin-select" popup-content-class="receive-coin-select-popup" @popup-show="addSearchField" v-model="destinationCoin" @filter="filterDestinationCoin" @input="updateCoinName()" :disabled="!destinationCoinOptions" :loading="!destinationCoinOptions" :options="destinationCoinOptions">
-                                                                <template v-slot:option="scope">
-                                                                    <q-item class="custom-menu" v-bind="scope.itemProps" v-on="scope.itemEvents">
-                                                                        <q-item-section avatar>
-                                                                            <q-icon class="option--avatar option--avatar__custom" :name="`img:${scope.opt.image}`" />
-                                                                        </q-item-section>
-                                                                        <q-item-section>
-                                                                            <q-item-label caption>{{ scope.opt.value }}</q-item-label>
-                                                                            <q-item-label v-html="scope.opt.label" />
-                                                                        </q-item-section>
-                                                                    </q-item>
-                                                                </template>
-                                                                <template v-slot:selected>
-                                                                    <q-item v-if="destinationCoin">
-                                                                        <q-item-section avatar>
-                                                                            <q-icon class="option--avatar option--avatar__custom" :name="`img:${destinationCoin.image}`" />
-                                                                        </q-item-section>
-                                                                        <q-item-section>
-                                                                            <q-item-label caption>{{ destinationCoin.value }}</q-item-label>
-                                                                            <q-item-label v-html="destinationCoin.label" />
-                                                                        </q-item-section>
-                                                                    </q-item>
-                                                                    <q-item v-else>
-                                                                    </q-item>
-                                                                </template>
-                                                            </q-select>
-                                                        </div>
-                                                        <div class="col col-2 choose-coin">
-                                                            <span class="cursor">
-                                                                <q-select class="select-input" light separator use-input borderless rounded v-model="depositCoin" @input="updateCoinName()" @filter="filterDepositCoin" :disabled="!depositCoinOptions" :loading="!depositCoinOptions" :options="depositCoinOptions">
-                                                                    <template v-slot:option="scope">
-                                                                        <q-item class="custom-menu" v-bind="scope.itemProps" v-on="scope.itemEvents">
-                                                                            <q-item-section avatar>
-                                                                                <q-icon :name="`img:${scope.opt.image}`" />
-                                                                            </q-item-section>
-                                                                            <q-item-section>
-                                                                                <q-item-label v-html="scope.opt.label" />
-                                                                                <q-item-label caption>{{ scope.opt.contract }}</q-item-label>
-                                                                                <q-item-label v-if="scope.opt.amount" caption>{{ scope.opt.amount }}</q-item-label>
-                                                                                <q-item-label v-if="scope.opt.name" caption>{{ scope.opt.name }}</q-item-label>
-                                                                            </q-item-section>
-                                                                        </q-item>
-                                                                    </template>
-                                                                    <template v-slot:selected>
-                                                                        <span class="text-h5 text-bold">{{depositCoin.value.toUpperCase()}}</span>
-                                                                        <q-item-label v-if="depositCoin.name" caption>{{ depositCoin.name }}</q-item-label>
-                                                                    </template>
-                                                                </q-select>
-
-                                                            </span>
-                                                        </div>
-                                                        <div class="col col-6 offset-4">
-                                                            <q-input outlined class="bg-white text-h5" ref="depositQuantity" @input="quantityFromDeposit()" v-model="depositQuantity" type="number" :disabled="!rateData" :loading="!rateData" :rules="[ val => val >= rateData.limitMinDepositCoin || 'This is less than the minimum allowed', val => val < rateData.limitMaxDepositCoin || 'This is more than the maximum allowed']">
-                                                                <div class="flex justify-end items-center" style="width: 60px">
-                                                                    <q-icon v-if="depositCoin" class="option--avatar" :name="`img:${depositCoin.image}`" />
-                                                                </div>
-                                                            </q-input>
-                                                        </div>
-                                                    </div>
-                                                </div>
-                                                <div class="you-receive">
-                                                    <br>
-                                                    <q-btn outline round color="black" icon="swap_vert" @click="switchAmounts()" class="swap_vert" />
-                                                    <div class="you-receive-head row items-center">
-                                                        <div class="col col-6">You Receive</div>
-                                                        <div v-if="rateData" class="col col-6 info_rate_holder small text-right flex justify-end items-center" :class="{'_loading': fetchingRate}">
-                                                            <!-- 1 ETH = 374.705 USDT -->
-                                                            <span>{{ '1 ' + fromCoinType.toUpperCase() + '&nbsp;= &nbsp;' + rateData.rate.toFixed(5) + ' ' + toCoinType.toUpperCase() }}</span>
-                                                            <div class="info_rate">
-                                                                <span class="i-btn">i
-                                                                    <!-- <img src="statics/theme1/rate-info.png" class="rate-info-prototype" alt=""> -->
-                                                                    <div class="info_rate_wrapper text-left">
-                                                                        <div class="title">We got you the best price</div>
-                                                                        <hr>
-                                                                        <div class="lineheight flex justify-between items-center">
-                                                                            <span>Max Deposit</span>
-                                                                            <span>{{rateData.limitMaxDepositCoin}} {{fromCoinType.toUpperCase()}}</span>
-                                                                        </div>
-                                                                        <div class="lineheight flex justify-between items-center">
-                                                                            <span>Max Destination</span>
-                                                                            <span>{{rateData.limitMaxDestinationCoin}} {{toCoinType.toUpperCase()}}</span>
-                                                                        </div>
-                                                                        <div class="lineheight flex justify-between items-center">
-                                                                            <span>Min Deposit</span>
-                                                                            <span>{{rateData.limitMinDepositCoin}} {{fromCoinType.toUpperCase()}}</span>
-                                                                        </div>
-                                                                        <div class="lineheight flex justify-between items-center">
-                                                                            <span>Min Destination</span>
-                                                                            <span>{{rateData.limitMinDestinationCoin}} {{toCoinType.toUpperCase()}}</span>
-                                                                        </div>
-                                                                    </div>
-                                                                </span>
-                                                            </div>
-                                                        </div>
-                                                    </div>
-                                                    <div class="you-receive-body row items-center">
-                                                        <div class="col col-2 choose-coin"><span class="cursor">
-
-                                                                <q-select class="select-input" light separator use-input borderless rounded v-model="destinationCoin" @input="updateCoinName()" @filter="filterDestinationCoin" :disabled="!destinationCoinOptions" :loading="!destinationCoinOptions" :options="destinationCoinOptions">
-                                                                    <template v-slot:option="scope">
-                                                                        <q-item class="custom-menu" v-bind="scope.itemProps" v-on="scope.itemEvents">
-                                                                            <q-item-section avatar>
-                                                                                <q-icon :name="`img:${scope.opt.image}`" />
-                                                                            </q-item-section>
-                                                                            <q-item-section>
-                                                                                <q-item-label v-html="scope.opt.label" />
-                                                                                <q-item-label caption>{{ scope.opt.contract }}</q-item-label>
-                                                                                <q-item-label v-if="scope.opt.amount" caption>{{ scope.opt.amount }}</q-item-label>
-                                                                                <q-item-label v-if="scope.opt.name" caption>{{ scope.opt.name }}</q-item-label>
-                                                                            </q-item-section>
-                                                                        </q-item>
-                                                                    </template>
-                                                                    <template v-slot:selected>
-                                                                        <span class="text-h5 text-bold">{{destinationCoin.value.toUpperCase()}}</span>
-                                                                        <q-item-label v-if="destinationCoin.name" caption>{{ destinationCoin.name }}</q-item-label>
-
-                                                                    </template>
-                                                                </q-select>
-                                                            </span></div>
-                                                        <div class="col col-6 offset-4">
-                                                            <q-input outlined class="bg-white text-h5" ref="destinationQuantity" v-model="destinationQuantity" @input="quantityFromDestination()" :disabled="!rateData" :loading="!rateData" type="number" :rules="[ val => val >= rateData.limitMinDestinationCoin || 'This is less than the minimum allowed', val => val < rateData.limitMaxDestinationCoin || 'This is more than the maximum allowed']">
-                                                                <div class="flex justify-end items-center" style="width: 60px">
-                                                                    <q-icon v-if="destinationCoin" class="option--avatar" :name="`img:${destinationCoin.image}`" />
-                                                                </div>
-                                                            </q-input>
-                                                        </div>
-                                                    </div>
-                                                    <q-btn unelevated @click="pStep = 2" color="light-grey" text-color="black" label="Choose Accounts" class="text-capitalize chose_accounts full-width" />
-                                                </div>
-                                            </div>
-                                            <div v-if="pStep === 2" class="prototype">
-                                                <div class="head">
-                                                    <q-btn flat @click="pStep = 1" unelevated icon="keyboard_arrow_left" rounded color="grey" label="Back" class="--next-btn q-mr-md" />
-                                                    Choose Accounts
-                                                </div>
-                                                <div class="you-pay">
-                                                    <div class="you-pay-head row items-center">
-                                                        <div class="col col-6">From</div>
-                                                        <!-- <div class="col col-6 red text-right text-red">Max 0 USDT</div> -->
-                                                    </div>
-                                                    <div class="you-pay-body row items-center">
-                                                        <div class="col col-12 pay-coin">
-                                                            <!-- class="select-input" -->
-                                                            <q-select v-show="fromCoin === null || (fromCoin.type !== 'new_public_key')" light separator outlined class="select-input" @input="checkGetPairs()" v-model="fromCoin" :options="optionsFrom">
-                                                                <template v-slot:option="scope">
-                                                                    <q-item class="custom-menu" v-bind="scope.itemProps" v-on="scope.itemEvents">
-                                                                        <q-item-section avatar>
-                                                                            <q-icon class="option--avatar option--avatar" :class="scope.opt.value" :name="`img:${scope.opt.image}`" />
-                                                                        </q-item-section>
-                                                                        <q-item-section dark>
-                                                                            <q-item-label v-html="scope.opt.label" />
-                                                                            <q-item-label caption class="ellipsis mw160">{{ scope.opt.value }}</q-item-label>
-                                                                        </q-item-section>
-                                                                    </q-item>
-                                                                </template>
-                                                                <template v-slot:selected>
-                                                                    <q-item v-if="fromCoin">
-                                                                        <q-item-section avatar>
-                                                                            <q-icon class="option--avatar option--avatar__custom" :name="`img:${fromCoin.image}`" />
-                                                                        </q-item-section>
-                                                                        <q-item-section>
-                                                                            <q-item-label v-html="fromCoin.label" />
-                                                                            <q-item-label caption class="ellipsis mw160">{{ fromCoin.value }}</q-item-label>
-                                                                        </q-item-section>
-                                                                    </q-item>
-                                                                    <q-item v-else>
-                                                                    </q-item>
-                                                                </template>
-                                                            </q-select>
-                                                            <q-input v-show="fromCoin !== null && (fromCoin.type === 'new_public_key')" v-model="refundAddress.address" @input="verifyAddress()" class="input-input" :label="depositCoin !== null ? 'Your ' + depositCoin.value.toUpperCase() + ' return address' : ' Your return address'" outlined color="purple" type="text">
-                                                                <template v-slot:append>
-                                                                    <div class="flex justify-end items-center">
-                                                                        <q-btn color="purple" rounded class="q-mb-xs" @click="fromCoin = null" outlined unelevated flat text-color="black" label="Hide" />
-                                                                    </div>
-                                                                </template>
-                                                            </q-input>
-                                                        </div>
-                                                        <div class="col col-12">
-                                                            <q-btn v-show="!fromCoinMemo" flat class="q-mt-sm q-mb-sm --next-btn align-left full-width" :icon-right="fromCoinMemo ? 'close':'add'" :label="fromCoinMemo ? 'Hide Tag/Memo':'Add Tag/Memo'" @click="fromCoinMemo = !fromCoinMemo" />
-                                                            <hr v-show="fromCoinMemo" style="opacity:0; height: 5px;margin: 0px">
-                                                            <q-input v-show="fromCoinMemo" class="input-input" outlined color="purple" type="text" v-model="refundAddress.tag" label="Optional tag or memo" hint="some exchanges require this field">
-                                                                <template v-slot:append>
-                                                                    <div class="flex justify-end items-center">
-                                                                        <q-btn color="purple" rounded class="q-mb-xs" @click="fromCoinMemo = false" outlined unelevated flat text-color="black" label="Hide" />
-                                                                    </div>
-                                                                </template>
-                                                            </q-input>
-                                                        </div>
-                                                    </div>
-                                                    <!-- <br> -->
-                                                </div>
-                                                <div class="you-receive">
-                                                    <div class="you-receive-head row items-center">
-                                                        <div class="col col-6">To</div>
-                                                    </div>
-                                                    <div class="you-receive-body row items-center">
-                                                        <div class="col col-12">
-                                                            <q-select v-show="toCoin === null || (toCoin.type !== 'new_public_key')" light separator outlined class="select-input" v-model="toCoin" @input="updateCoinName()" use-input :options="optionsTo">
-                                                                <template v-slot:option="scope">
-                                                                    <q-item class="custom-menu" v-bind="scope.itemProps" v-on="scope.itemEvents">
-                                                                        <q-item-section avatar>
-                                                                            <q-icon class="option--avatar option--avatar" :class="scope.opt.value" :name="`img:${scope.opt.image}`" />
-                                                                        </q-item-section>
-                                                                        <q-item-section dark>
-                                                                            <q-item-label v-html="scope.opt.label" />
-                                                                            <q-item-label caption class="ellipsis mw160">{{ scope.opt.value }}</q-item-label>
-                                                                        </q-item-section>
-                                                                    </q-item>
-                                                                </template>
-                                                                <template v-slot:selected>
-                                                                    <q-item v-if="toCoin">
-                                                                        <q-item-section avatar>
-                                                                            <q-icon class="option--avatar option--avatar__custom" :class="toCoin.value" :name="`img:${toCoin.image}`" />
-                                                                        </q-item-section>
-                                                                        <q-item-section>
-                                                                            <q-item-label v-html="toCoin.label" />
-                                                                            <q-item-label caption class="ellipsis mw160">{{ toCoin.value }}</q-item-label>
-                                                                        </q-item-section>
-                                                                    </q-item>
-                                                                    <q-item v-else>
-                                                                    </q-item>
-                                                                </template>
-                                                            </q-select>
-                                                        </div>
-                                                        <div class="col col-12">
-                                                            <!-- :rules="[ val => val.length >= 3 || 'Destination Address Cannot less than 3 characters' ]" -->
-                                                            <q-input v-show="toCoin !== null && toCoin.type === 'new_public_key'" ref="destinationAddressAddress" :label="destinationAddressLabel" v-model="destinationAddress.address" @input="verifyAddress()" class="input-input destinationAddressAddress" outlined color="purple" type="text">
-                                                                <template v-slot:append>
-                                                                    <div class="flex justify-end">
-                                                                        <q-btn color="purple" rounded class="q-mb-sm" @click="toCoin = null" outlined unelevated flat text-color="black" label="Hide" />
-                                                                    </div>
-                                                                </template>
-                                                            </q-input>
-                                                            <hr v-show="toCoinMemo" style="opacity:0; height: 5px;margin: 0px">
-                                                            <q-input v-show="toCoinMemo" class="input-input" outlined color="purple" type="text" v-model="destinationAddress.tag" label="Optional tag or memo" hint="some exchanges require this field">
-                                                                <template v-slot:append>
-                                                                    <div class="flex justify-end">
-                                                                        <q-btn color="purple" rounded class="q-mb-sm" @click="toCoinMemo = false" outlined unelevated flat text-color="black" label="Hide" />
-                                                                    </div>
-                                                                </template>
-                                                            </q-input>
-                                                            <q-btn v-show="!toCoinMemo" flat class="q-mt-sm q-mb-sm --next-btn align-left full-width" :icon-right="toCoinMemo ? 'close':'add'" :label="toCoinMemo ? 'Hide Tag/Memo':'Add Tag/Memo'" @click="toCoinMemo = !toCoinMemo" />
-                                                            <hr v-show="toCoinMemo" style="opacity:0; height: 35px;margin: 0px">
-                                                        </div>
-                                                        <div class="col col-12 flex justify-end disclaimer-wrapper">
-                                                            <div class="row full-width">
-                                                                <div class="q-gutter-sm">
-                                                                    <q-btn color="white" flat @click="showDisclaimerWrapper = true" class="lower bold" text-color="black" label="Read the disclaimer" />
-                                                                </div>
-                                                            </div>
-                                                            <div class="row full-width" style="padding-left: 13px; margin-top: 10px;">
-                                                                <div class="q-gutter-sm">
-                                                                    <q-checkbox label="I accept" color="deep-purple-14" v-model="disclaimerCheck" />
-                                                                </div>
-                                                            </div>
-                                                        </div>
-                                                    </div>
-                                                    <br>
-                                                    <q-btn unelevated @click="checkAddressMatchCoins()" :disable="!disclaimerCheck" color="light-grey" text-color="black" label="Review Order" class="text-capitalize chose_accounts full-width" />
-                                                </div>
-                                            </div>
-                                            <div v-if="pStep === 3" class="prototype">
-                                                <div class="head">
-                                                    <q-btn flat @click="pStep = 2" unelevated icon="keyboard_arrow_left" rounded color="grey" label="Back" class="--next-btn q-mr-md" />
-                                                    Order in progress
-                                                </div>
-                                                <div class="standard-content--body">
-                                                    <div class="standard-content--body__form q-pa-xl">
-                                                        <div class="progress-custom-volentix column flex-center">
-                                                            <svg class="svg_logo" fill="#7272FA" width="40" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 16 20.58">
-                                                                <path d="M199,25.24q0,3.29,0,6.57a.5.5,0,0,1-.18.41l-7.32,6.45a.57.57,0,0,1-.71,0l-7.21-6.1c-.12-.11-.25-.22-.38-.32a.53.53,0,0,1-.22-.47q0-3.83,0-7.66,0-2.69,0-5.39c0-.33.08-.47.29-.51s.33.07.44.37l3.45,8.84c.52,1.33,1,2.65,1.56,4a.21.21,0,0,0,.23.16h4.26a.19.19,0,0,0,.21-.14l3.64-9.7,1.21-3.22c.08-.22.24-.32.42-.29a.34.34,0,0,1,.27.37c0,.41,0,.81,0,1.22Q199,22.53,199,25.24Zm-8.75,12s0,0,0,0,0,0,0,0a.28.28,0,0,0,0-.05l-1.88-4.83c0-.11-.11-.11-.2-.11h-3.69s-.1,0-.13,0l.11.09,4.48,3.8C189.38,36.55,189.8,36.93,190.25,37.27Zm-6.51-16.76h0s0,.07,0,.1q0,5.4,0,10.79c0,.11,0,.16.15.16h4.06c.15,0,.15,0,.1-.16s-.17-.44-.26-.66l-3.1-7.94Zm14.57.06c-.06,0-.06.07-.07.1l-1.89,5q-1.06,2.83-2.13,5.66c-.06.16,0,.19.13.19h3.77c.16,0,.2,0,.2-.2q0-5.3,0-10.59Zm-7.16,17,.05-.11,1.89-5c.05-.13,0-.15-.11-.15h-3.71c-.17,0-.16,0-.11.18.26.65.51,1.31.77,2Zm.87-.3,0,0,5.65-5H194c-.13,0-.16.07-.19.17l-1.59,4.23Zm0,.06h0Z" transform="translate(-183 -18.21)"></path>
-                                                            </svg>
-                                                            <span class="title">{{ friendlyStatus }}</span>
-                                                            <q-linear-progress indeterminate stripe rounded size="md" :value="progress" class="q-mt-md" />
-                                                        </div>
-                                                        <hr style="height:15px;opacity:0" />
-                                                        <div class="text-black">
-                                                            <div class="text-h4 --subtitle">{{exchangeLabel}}</div>
-                                                            <q-input v-model="exchangeAddress.address" readonly rounded class="input-input pr80" outlined color="purple" type="text">
-                                                                <template v-slot:append>
-                                                                    <div class="flex justify-end">
-                                                                        <q-btn flat unelevated text-color="grey" @click="copyToClipboard(exchangeAddress.address , 'Exchange Address')" round class="btn-copy" icon="file_copy" />
-                                                                    </div>
-                                                                </template>
-                                                            </q-input>
-                                                            <br v-if="exchangeAddress.tag">
-                                                            <q-input v-if="exchangeAddress.tag" v-model="exchangeAddress.tag" readonly rounded class="input-input pr80" outlined color="purple" type="text">
-                                                                <template v-slot:append>
-                                                                    <div class="flex justify-end">
-                                                                        <q-btn flat unelevated text-color="grey" @click="copyToClipboard(exchangeAddress.tag , 'Exchange Tag')" round class="btn-copy" icon="file_copy" />
-                                                                    </div>
-                                                                </template>
-                                                            </q-input>
-                                                        </div>
-                                                    </div>
-                                                </div>
-                                            </div>
-                                        </div>
-                                    </div>
-                                    <!-- <div class="col col-6">
-                      <img src="statics/theme1/Screenshot_208.png" alt="">
-                    </div> -->
-                                </div>
-                                <div v-show="false" class="list-wrapper--chain__eos-to-vtx-convertor">
-                                    <q-stepper v-model="step" done-color="green" ref="stepper" class="stepper--desktop" alternative-labels color="primary" animated flat>
-                                        <!-- 1. Select Coin to Send -->
-                                        <q-step default title="Select Coin to Send" :name="1" prefix="1" :done="step > 1">
-                                            <div class="text-black">
-                                                <span class="sublab-input">Step 1</span><span class="tlab-input">Select Coin to Send</span>
-                                                <q-select light separator rounded outlined class="select-input" v-model="depositCoin" use-input @filter="filterDepositCoin" @input="checkGetPairs()" :disabled="!depositCoinOptions" :loading="!depositCoinOptions" :options="depositCoinOptions">
-                                                    <template v-slot:option="scope">
-                                                        <q-item class="custom-menu" v-bind="scope.itemProps" v-on="scope.itemEvents">
-                                                            <q-item-section avatar>
-                                                                <q-icon class="option--avatar option--avatar__custom" :name="`img:${scope.opt.image}`" />
-                                                            </q-item-section>
-                                                            <q-item-section dark>
-                                                                <q-item-label v-html="scope.opt.label" />
-                                                                <q-item-label caption>{{ scope.opt.value }}</q-item-label>
-                                                            </q-item-section>
-                                                        </q-item>
-                                                    </template>
-                                                    <template v-slot:selected>
-                                                        <q-item v-if="depositCoin">
-                                                            <q-item-section avatar>
-                                                                <q-icon class="option--avatar option--avatar__custom" :name="`img:${depositCoin.image}`" />
-                                                            </q-item-section>
-                                                            <q-item-section>
-                                                                <q-item-label v-html="depositCoin.label" />
-                                                                <q-item-label caption>{{ depositCoin.value }}</q-item-label>
-                                                            </q-item-section>
-                                                        </q-item>
-                                                        <q-item v-else>
-                                                        </q-item>
-                                                    </template>
-                                                </q-select>
-                                                <q-stepper-navigation v-show="true" class="flex justify-end">
-                                                    <q-btn @click="checkToGetPairs()" color="deep-purple-14" class="--next-btn" rounded :label="$t('next')" />
-                                                </q-stepper-navigation>
-                                            </div>
-                                        </q-step>
-                                        <!-- 2. Select Coin to Receive -->
-                                        <q-step default title="Select Coin to Receive" :name="2" prefix="2" :done="step > 2">
-                                            <!-- <q-btn flat @click="$refs.stepper.previous()" unelevated icon="keyboard_arrow_left" color="primary" label="Back" class="--back-btn"/> -->
-                                            <span class="sublab-input">Step 2</span><span class="tlab-input">Select Coin to Receive</span>
-                                            <div class="text-black">
-                                                <!-- <span class="lab-input">Select Coin to receive</span> -->
-                                                <q-select light separator rounded outlined class="select-input" v-model="destinationCoin" use-input @filter="filterDestinationCoin" @input="updateCoinName()" :disabled="!destinationCoinOptions" :loading="!destinationCoinOptions" :options="destinationCoinOptions">
-                                                    <template v-slot:option="scope">
-                                                        <q-item class="custom-menu" v-bind="scope.itemProps" v-on="scope.itemEvents">
-                                                            <q-item-section avatar>
-                                                                <q-icon class="option--avatar option--avatar__custom" :name="`img:${scope.opt.image}`" />
-                                                            </q-item-section>
-                                                            <q-item-section>
-                                                                <q-item-label v-html="scope.opt.label" />
-                                                                <q-item-label caption>{{ scope.opt.value }}</q-item-label>
-                                                            </q-item-section>
-                                                        </q-item>
-                                                    </template>
-                                                    <template v-slot:selected>
-                                                        <q-item v-if="destinationCoin">
-                                                            <q-item-section avatar>
-                                                                <q-icon class="option--avatar option--avatar__custom" :name="`img:${destinationCoin.image}`" />
-                                                            </q-item-section>
-                                                            <q-item-section>
-                                                                <q-item-label v-html="destinationCoin.label" />
-                                                                <q-item-label caption>{{ destinationCoin.value }}</q-item-label>
-                                                            </q-item-section>
-                                                        </q-item>
-                                                        <q-item v-else>
-                                                        </q-item>
-                                                    </template>
-                                                </q-select>
-                                                <q-stepper-navigation v-show="true" class="flex justify-end">
-                                                    <q-btn flat @click="$refs.stepper.previous()" unelevated icon="keyboard_arrow_left" rounded color="grey" label="Back" class="--next-btn q-mr-md" />
-                                                    <q-btn @click="checkToGetRate()" color="deep-purple-14" class="--next-btn" rounded :label="$t('next')" />
-                                                </q-stepper-navigation>
-                                            </div>
-                                        </q-step>
-                                        <!-- 3. Select Quantity -->
-                                        <q-step default title="Select Quantity" :name="3" prefix="3" :done="step > 3">
-                                            <!-- <q-btn flat @click="$refs.stepper.previous()" unelevated icon="keyboard_arrow_left" color="primary" label="Back" class="--back-btn"/> -->
-                                            <span class="sublab-input">Step 3</span><span class="tlab-input">Select Quantity</span>
-                                            <div class="standard-content--body q-pl-lg">
-                                                <div class="standard-content--body__form">
-                                                    <div class="pay-get-wrapper column justify-between">
-                                                        <div class="pay-wrapper column">
-                                                            <span class="label">You Send</span>
-                                                            <span class="value">
-                                                                <q-input ref="depositQuantity" @input="quantityFromDeposit()" rounded class="full-width pl0" flat v-model="depositQuantity" type="number" :disabled="!rateData" :loading="!rateData" :rules="[ val => val >= rateData.limitMinDepositCoin || 'This is less than the minimum allowed', val => val < rateData.limitMaxDepositCoin || 'This is more than the maximum allowed']">
-                                                                    <div class="flex justify-end items-center" style="width: 140px">
-                                                                        <q-icon v-if="depositCoin" class="option--avatar" :name="`img:${depositCoin.image}`" />
-                                                                        <span class="q-pl-sm text-bold text-h6">{{fromCoinType.toUpperCase()}}</span>
-                                                                    </div>
-                                                                </q-input>
-                                                            </span>
-                                                        </div>
-                                                        <q-btn flat unelevated class="exchange-btn" @click="switchAmounts()" text-color="black">
-                                                            <q-icon name="keyboard_backspace" class="left-icon" />
-                                                            <q-icon name="keyboard_backspace" class="right-icon" />
-                                                        </q-btn>
-                                                        <div class="get-wrapper column">
-                                                            <span class="label">You Receive</span>
-                                                            <span class="value">
-                                                                <q-input rounded class="full-width pl0" flat ref="destinationQuantity" v-model="destinationQuantity" @input="quantityFromDestination()" :disabled="!rateData" :loading="!rateData" :rules="[ val => val >= rateData.limitMinDestinationCoin || 'This is less than the minimum allowed', val => val < rateData.limitMaxDestinationCoin || 'This is more than the maximum allowed']" type="number">
-                                                                    <div class="flex justify-end items-center" style="width: 140px">
-                                                                        <q-icon v-if="destinationCoin" class="option--avatar" :name="`img:${destinationCoin.image}`" />
-                                                                        <span class="q-pl-sm text-bold text-h6">{{toCoinType.toUpperCase()}}</span>
-                                                                    </div>
-                                                                </q-input>
-                                                            </span>
-                                                        </div>
-                                                    </div>
-                                                    <div class="rate-value flex justify-end">
-                                                        <span class="label">Rate {{fromCoinType.toUpperCase() + '&nbsp;'}} </span>
-                                                        <span class="value"> = {{ rateData !== null ? rateData.rate : '0.03254'}} {{toCoinType.toUpperCase()}}</span>
-                                                    </div>
-                                                    <br>
-                                                </div>
-                                                <q-stepper-navigation v-show="true" class="flex justify-end">
-                                                    <q-btn flat @click="$refs.stepper.previous()" unelevated icon="keyboard_arrow_left" rounded color="grey" label="Back" class="--next-btn q-mr-md" />
-                                                    <q-btn @click="checkToGetRate()" color="deep-purple-14" class="--next-btn" rounded :label="$t('next')" />
-                                                </q-stepper-navigation>
-                                            </div>
-                                        </q-step>
-                                        <!-- 4. Chose account -->
-                                        <q-step default title="Chose account" :name="4" prefix="4" :done="step > 4">
-                                            <!-- <q-btn flat @click="$refs.stepper.previous()" unelevated icon="keyboard_arrow_left" color="primary" label="Back" class="--back-btn"/> -->
-                                            <span class="sublab-input">Step 4</span><span class="tlab-input">Chose account</span>
-                                            <div class="standard-content--body">
-                                                <div class="standard-content--body__form">
-                                                    <div class="row">
-                                                        <div class="col col-6 q-pr-md">
-                                                            <span v-show="fromCoin === null || (fromCoin.type !== 'new_public_key')" class="lab-input">From</span>
-                                                            <q-select v-show="fromCoin === null || (fromCoin.type !== 'new_public_key')" light separator rounded outlined class="select-input" @input="checkGetPairs()" v-model="fromCoin" :options="optionsFrom">
-                                                                <template v-slot:option="scope">
-                                                                    <q-item class="custom-menu" v-bind="scope.itemProps" v-on="scope.itemEvents">
-                                                                        <q-item-section avatar>
-                                                                            <q-icon class="option--avatar option--avatar" :class="scope.opt.value" :name="`img:${scope.opt.image}`" />
-                                                                        </q-item-section>
-                                                                        <q-item-section dark>
-                                                                            <q-item-label v-html="scope.opt.label" />
-                                                                            <q-item-label caption class="ellipsis mw160">{{ scope.opt.value }}</q-item-label>
-                                                                        </q-item-section>
-                                                                    </q-item>
-                                                                </template>
-                                                                <template v-slot:selected>
-                                                                    <q-item v-if="fromCoin">
-                                                                        <q-item-section avatar>
-                                                                            <q-icon class="option--avatar option--avatar__custom" :name="`img:${fromCoin.image}`" />
-                                                                        </q-item-section>
-                                                                        <q-item-section>
-                                                                            <q-item-label v-html="fromCoin.label" />
-                                                                            <q-item-label caption class="ellipsis mw160">{{ fromCoin.value }}</q-item-label>
-                                                                        </q-item-section>
-                                                                    </q-item>
-                                                                    <q-item v-else>
-                                                                    </q-item>
-                                                                </template>
-                                                            </q-select>
-                                                            <span v-show="fromCoin !== null && (fromCoin.type === 'new_public_key')" class="lab-input">Your <strong>{{ depositCoin !== null ? depositCoin.value.toUpperCase() : '' }}</strong> return address </span>
-                                                            <q-input v-show="fromCoin !== null && (fromCoin.type === 'new_public_key')" v-model="refundAddress.address" @input="verifyAddress()" class="input-input" rounded outlined color="purple" type="text" hint="[ in case the transaction does not complete ]">
-                                                                <template v-slot:append>
-                                                                    <div class="flex justify-end">
-                                                                        <q-btn color="purple" rounded class="q-mb-sm" @click="fromCoin = null" outlined unelevated flat text-color="black" label="Hide" />
-                                                                    </div>
-                                                                </template>
-                                                            </q-input>
-                                                            <br v-show="fromCoin !== null && (fromCoin.type === 'new_public_key')">
-                                                            <span class="lab-input" v-show="fromCoinMemo" />
-                                                            <q-input v-show="fromCoinMemo" class="input-input" rounded outlined color="purple" type="text" v-model="refundAddress.tag" label="Optional tag or memo" hint="some exchanges require this field">
-                                                                <template v-slot:append>
-                                                                    <div class="flex justify-end">
-                                                                        <q-btn color="purple" rounded class="q-mb-sm" @click="fromCoinMemo = false" outlined unelevated flat text-color="black" label="Hide" />
-                                                                    </div>
-                                                                </template>
-                                                            </q-input>
-                                                            <br v-show="fromCoinMemo">
-                                                            <q-btn v-show="!fromCoinMemo" flat class="q-mt-sm q-mb-sm --next-btn align-left full-width" :icon-right="fromCoinMemo ? 'close':'add'" rounded :label="fromCoinMemo ? 'Hide Tag/Memo':'Add Tag/Memo'" @click="fromCoinMemo = !fromCoinMemo" />
-                                                            <br>
-                                                        </div>
-                                                        <div class="col col-6">
-                                                            <!-- <hr> -->
-                                                            <span v-show="toCoin === null || (toCoin.type !== 'new_public_key')" class="lab-input">To</span>
-                                                            <q-select v-show="toCoin === null || (toCoin.type !== 'new_public_key')" light separator rounded outlined class="select-input" v-model="toCoin" @input="updateCoinName()" use-input :options="optionsTo">
-                                                                <template v-slot:option="scope">
-                                                                    <q-item class="custom-menu" v-bind="scope.itemProps" v-on="scope.itemEvents">
-                                                                        <q-item-section avatar>
-                                                                            <q-icon class="option--avatar option--avatar" :class="scope.opt.value" :name="`img:${scope.opt.image}`" />
-                                                                        </q-item-section>
-                                                                        <q-item-section dark>
-                                                                            <q-item-label v-html="scope.opt.label" />
-                                                                            <q-item-label caption class="ellipsis mw160">{{ scope.opt.value }}</q-item-label>
-                                                                        </q-item-section>
-                                                                    </q-item>
-                                                                </template>
-                                                                <template v-slot:selected>
-                                                                    <q-item v-if="toCoin">
-                                                                        <q-item-section avatar>
-                                                                            <q-icon class="option--avatar option--avatar__custom" :class="toCoin.value" :name="`img:${toCoin.image}`" />
-                                                                        </q-item-section>
-                                                                        <q-item-section>
-                                                                            <q-item-label v-html="toCoin.label" />
-                                                                            <q-item-label caption class="ellipsis mw160">{{ toCoin.value }}</q-item-label>
-                                                                        </q-item-section>
-                                                                    </q-item>
-                                                                    <q-item v-else>
-                                                                    </q-item>
-                                                                </template>
-                                                            </q-select>
-                                                            <span v-show="toCoin !== null && toCoin.type === 'new_public_key'" class="lab-input">{{ destinationAddressLabel }}</span>
-                                                            <!-- :rules="[ val => val.length >= 3 || 'Destination Address Cannot less than 3 characters' ]" -->
-                                                            <q-input v-show="toCoin !== null && toCoin.type === 'new_public_key'" ref="destinationAddressAddress" v-model="destinationAddress.address" @input="verifyAddress()" class="input-input destinationAddressAddress" rounded outlined color="purple" type="text">
-                                                                <template v-slot:append>
-                                                                    <div class="flex justify-end">
-                                                                        <q-btn color="purple" rounded class="q-mb-sm" @click="toCoin = null" outlined unelevated flat text-color="black" label="Hide" />
-                                                                    </div>
-                                                                </template>
-                                                            </q-input>
-                                                            <!-- <br v-show="toCoin !== null && (toCoin.type === 'new_public_key')"> -->
-                                                            <span class="lab-input" v-show="toCoinMemo" :class="{'margbotm' : toCoinMemo}" />
-                                                            <q-input v-show="toCoinMemo" class="input-input" rounded outlined color="purple" type="text" v-model="destinationAddress.tag" label="Optional tag or memo" hint="some exchanges require this field">
-                                                                <template v-slot:append>
-                                                                    <div class="flex justify-end">
-                                                                        <q-btn color="purple" rounded class="q-mb-sm" @click="toCoinMemo = false" outlined unelevated flat text-color="black" label="Hide" />
-                                                                    </div>
-                                                                </template>
-                                                            </q-input>
-                                                            <br v-show="toCoinMemo">
-                                                            <q-btn v-show="!toCoinMemo" flat class="q-mt-sm q-mb-sm --next-btn align-left full-width" :icon-right="toCoinMemo ? 'close':'add'" rounded :label="toCoinMemo ? 'Hide Tag/Memo':'Add Tag/Memo'" @click="toCoinMemo = !toCoinMemo" />
-                                                        </div>
-                                                    </div>
-                                                </div>
-                                            </div>
-                                            <q-stepper-navigation v-show="true" class="flex justify-end">
-                                                <div class="row full-width" style="padding-left: 6px; margin-top: -20px;">
-                                                    <div class="q-gutter-sm">
-                                                        <q-btn color="white" flat @click="showDisclaimerWrapper = true" class="lower bold" text-color="black" label="Read the disclaimer" />
-                                                    </div>
-                                                </div>
-                                                <div class="row full-width" style="padding-left: 13px; margin-top: 10px;">
-                                                    <div class="q-gutter-sm">
-                                                        <q-checkbox label="I accept" color="deep-purple-14" v-model="disclaimerCheck" />
-                                                    </div>
-                                                </div>
-                                                <div class="standard-content--footer">
-                                                    <q-btn flat @click="$refs.stepper.previous()" unelevated icon="keyboard_arrow_left" rounded color="grey" label="Back" class="--next-btn q-mr-md" />
-                                                    <q-btn @click="checkAddressMatchCoins()" :disable="!disclaimerCheck" flat class="action-link next" color="black" text-color="white">
-                                                        <span class="label">Exchange {{ fromCoinType.toUpperCase() }}
-                                                            <q-icon name="keyboard_backspace" color="white" class="left-icon" /> {{ toCoinType.toUpperCase() }}
-                                                        </span>
-                                                    </q-btn>
-                                                </div>
-                                            </q-stepper-navigation>
-                                        </q-step>
-                                        <!-- 5. View Order -->
-                                        <q-step default title="View Order" :name="5" prefix="5" :done="step > 5">
-                                            <q-btn flat @click="$refs.stepper.previous()" unelevated icon="keyboard_arrow_left" color="primary" label="Back" class="--back-btn" />
-                                            <div class="standard-content--body">
-                                                <div class="standard-content--body__form" style="margin-left: -35px;">
-                                                    <div class="progress-custom-volentix column flex-center">
-                                                        <svg class="svg_logo" fill="#7272FA" width="40" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 16 20.58">
-                                                            <path d="M199,25.24q0,3.29,0,6.57a.5.5,0,0,1-.18.41l-7.32,6.45a.57.57,0,0,1-.71,0l-7.21-6.1c-.12-.11-.25-.22-.38-.32a.53.53,0,0,1-.22-.47q0-3.83,0-7.66,0-2.69,0-5.39c0-.33.08-.47.29-.51s.33.07.44.37l3.45,8.84c.52,1.33,1,2.65,1.56,4a.21.21,0,0,0,.23.16h4.26a.19.19,0,0,0,.21-.14l3.64-9.7,1.21-3.22c.08-.22.24-.32.42-.29a.34.34,0,0,1,.27.37c0,.41,0,.81,0,1.22Q199,22.53,199,25.24Zm-8.75,12s0,0,0,0,0,0,0,0a.28.28,0,0,0,0-.05l-1.88-4.83c0-.11-.11-.11-.2-.11h-3.69s-.1,0-.13,0l.11.09,4.48,3.8C189.38,36.55,189.8,36.93,190.25,37.27Zm-6.51-16.76h0s0,.07,0,.1q0,5.4,0,10.79c0,.11,0,.16.15.16h4.06c.15,0,.15,0,.1-.16s-.17-.44-.26-.66l-3.1-7.94Zm14.57.06c-.06,0-.06.07-.07.1l-1.89,5q-1.06,2.83-2.13,5.66c-.06.16,0,.19.13.19h3.77c.16,0,.2,0,.2-.2q0-5.3,0-10.59Zm-7.16,17,.05-.11,1.89-5c.05-.13,0-.15-.11-.15h-3.71c-.17,0-.16,0-.11.18.26.65.51,1.31.77,2Zm.87-.3,0,0,5.65-5H194c-.13,0-.16.07-.19.17l-1.59,4.23Zm0,.06h0Z" transform="translate(-183 -18.21)"></path>
-                                                        </svg>
-                                                        <span class="title">{{ friendlyStatus }}</span>
-                                                        <q-linear-progress indeterminate stripe rounded size="md" :value="progress" class="q-mt-md" />
-                                                    </div>
-                                                    <hr style="height:15px;opacity:0" />
-                                                    <div class="text-black">
-                                                        <div class="text-h4 --subtitle">
-                                                            <ul>
-                                                                <li><span>{{exchangeLabel}}</span></li>
-                                                            </ul>
-                                                        </div>
-                                                        <q-input v-model="exchangeAddress.address" readonly rounded class="input-input pr80" outlined color="purple" type="text">
-                                                            <template v-slot:append>
-                                                                <div class="flex justify-end">
-                                                                    <q-btn flat unelevated text-color="grey" @click="copyToClipboard(exchangeAddress.address , 'Exchange Address')" round class="btn-copy" icon="o_file_copy" />
-                                                                </div>
-                                                            </template>
-                                                        </q-input>
-                                                    </div>
-                                                </div>
-                                            </div>
-                                        </q-step>
-                                    </q-stepper>
-                                </div>
-                                <br><br><br>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </div>
-    <div v-else class="chain-tools-wrapper">
-        <div class="standard-content">
-            <h2 class="standard-content--title flex justify-center">
-                <q-btn flat unelevated class="btn-align-left" to="goBack" text-color="black" icon="keyboard_backspace" />
-                Exchange
-            </h2>
-            <div class="exchange_picto flex flex-center"><img src="statics/exchange_picto.svg" alt=""></div>
-        </div>
-        <div class="chain-tools-wrapper--list open">
-            <div class="list-wrapper">
-                <div class="list-wrapper--chain__eos-to-vtx-convertor">
-                    <q-stepper v-model="step" done-color="green" ref="stepper" alternative-labels vertical color="primary" animated flat>
-                        <!-- 1. Select Coin to Send -->
-                        <q-step default title="Select Coin to Send" :name="1" prefix="1" :done="step > 1">
-                            <div class="text-black">
-                                <!-- <span class="lab-input">Select Coin to Send</span> -->
-                                <q-select light separator rounded outlined class="select-input" v-model="depositCoin" use-input @filter="filterDepositCoin" @input="checkGetPairs()" :disabled="!depositCoinOptions" :loading="!depositCoinOptions" :options="depositCoinOptions">
-                                    <template v-slot:option="scope">
-                                        <q-item class="custom-menu" v-bind="scope.itemProps" v-on="scope.itemEvents">
-                                            <q-item-section avatar>
-                                                <q-icon class="option--avatar option--avatar__custom" :name="`img:${scope.opt.image}`" />
-                                            </q-item-section>
-                                            <q-item-section dark>
-                                                <q-item-label v-html="scope.opt.label" />
-                                                <q-item-label caption>{{ scope.opt.value }}</q-item-label>
-                                            </q-item-section>
-                                        </q-item>
-                                    </template>
-                                    <template v-slot:selected>
-                                        <q-item v-if="depositCoin">
-                                            <q-item-section avatar>
-                                                <q-icon class="option--avatar option--avatar__custom" :name="`img:${depositCoin.image}`" />
-                                            </q-item-section>
-                                            <q-item-section>
-                                                <q-item-label v-html="depositCoin.label" />
-                                                <q-item-label caption>{{ depositCoin.value }}</q-item-label>
-                                            </q-item-section>
-                                        </q-item>
-                                        <q-item v-else>
-                                        </q-item>
-                                    </template>
-                                </q-select>
-                                <q-stepper-navigation v-show="true" class="flex justify-end">
-                                    <q-btn @click="checkToGetPairs()" color="deep-purple-14" class="--next-btn" rounded :label="$t('next')" />
-                                </q-stepper-navigation>
-                            </div>
-                        </q-step>
-                        <!-- 2. Select Coin to Receive -->
-                        <q-step default title="Select Coin to Receive" :name="2" prefix="2" :done="step > 2">
-                            <q-btn flat @click="$refs.stepper.previous()" unelevated icon="keyboard_arrow_left" color="primary" label="Back" class="--back-btn" />
-                            <div class="text-black">
-                                <!-- <span class="lab-input">Select Coin to receive</span> -->
-                                <q-select light separator rounded outlined class="select-input" v-model="destinationCoin" use-input @filter="filterDestinationCoin" @input="updateCoinName()" :disabled="!destinationCoinOptions" :loading="!destinationCoinOptions" :options="destinationCoinOptions">
-                                    <template v-slot:option="scope">
-                                        <q-item class="custom-menu" v-bind="scope.itemProps" v-on="scope.itemEvents">
-                                            <q-item-section avatar>
-                                                <q-icon class="option--avatar option--avatar__custom" :name="`img:${scope.opt.image}`" />
-                                            </q-item-section>
-                                            <q-item-section>
-                                                <q-item-label v-html="scope.opt.label" />
-                                                <q-item-label caption>{{ scope.opt.value }}</q-item-label>
-                                            </q-item-section>
-                                        </q-item>
-                                    </template>
-                                    <template v-slot:selected>
-                                        <q-item v-if="destinationCoin">
-                                            <q-item-section avatar>
-                                                <q-icon class="option--avatar option--avatar__custom" :name="`img:${destinationCoin.image}`" />
-                                            </q-item-section>
-                                            <q-item-section>
-                                                <q-item-label v-html="destinationCoin.label" />
-                                                <q-item-label caption>{{ destinationCoin.value }}</q-item-label>
-                                            </q-item-section>
-                                        </q-item>
-                                        <q-item v-else>
-                                        </q-item>
-                                    </template>
-                                </q-select>
-                                <q-stepper-navigation v-show="true" class="flex justify-end">
-                                    <q-btn @click="checkToGetRate()" color="deep-purple-14" class="--next-btn" rounded :label="$t('next')" />
-                                </q-stepper-navigation>
-                            </div>
-                        </q-step>
-                        <!-- 3. Select Quantity -->
-                        <q-step default title="Select Quantity" :name="3" prefix="3" :done="step > 3">
-                            <q-btn flat @click="$refs.stepper.previous()" unelevated icon="keyboard_arrow_left" color="primary" label="Back" class="--back-btn" />
-                            <div class="standard-content--body">
-                                <div class="standard-content--body__form">
-                                    <div class="pay-get-wrapper column justify-between">
-                                        <div class="pay-wrapper column">
-                                            <span class="label">you pay</span>
-                                            <span class="value">
-                                                <q-input ref="depositQuantity" @input="quantityFromDeposit()" :suffix="fromCoinType.toUpperCase()" rounded class="full-width pl0" flat v-model="depositQuantity" type="number" :disabled="!rateData" :loading="!rateData" :rules="[ val => val >= rateData.limitMinDepositCoin || 'This is less than the minimum allowed', val => val < rateData.limitMaxDepositCoin || 'This is more than the maximum allowed']" />
-                                            </span>
-                                        </div>
-                                        <q-btn flat unelevated class="exchange-btn" @click="switchAmounts()" text-color="black">
-                                            <q-icon name="keyboard_backspace" class="left-icon" />
-                                            <q-icon name="keyboard_backspace" class="right-icon" />
-                                        </q-btn>
-                                        <div class="get-wrapper column">
-                                            <span class="label">you get</span>
-                                            <span class="value">
-                                                <q-input rounded class="full-width pl0" flat ref="destinationQuantity" :suffix="toCoinType.toUpperCase()" v-model="destinationQuantity" @input="quantityFromDestination()" :disabled="!rateData" :loading="!rateData" :rules="[ val => val >= rateData.limitMinDestinationCoin || 'This is less than the minimum allowed', val => val < rateData.limitMaxDestinationCoin || 'This is more than the maximum allowed']" type="number" />
-                                            </span>
-                                        </div>
-                                    </div>
-                                    <div class="rate-value flex justify-end">
-                                        <span class="label">Rate {{fromCoinType.toUpperCase() + '&nbsp;'}} </span>
-                                        <span class="value"> = {{ rateData !== null ? rateData.rate : '0.03254'}} {{toCoinType.toUpperCase()}}</span>
-                                    </div>
-                                    <br>
-                                </div>
-                                <q-stepper-navigation v-show="true" class="flex justify-end">
-                                    <q-btn @click="checkToGetRate()" color="deep-purple-14" class="--next-btn" rounded :label="$t('next')" />
-                                </q-stepper-navigation>
-                            </div>
-                        </q-step>
-                        <!-- 4. Chose account -->
-                        <q-step default title="Chose account" :name="4" prefix="4" :done="step > 4">
-                            <q-btn flat @click="$refs.stepper.previous()" unelevated icon="keyboard_arrow_left" color="primary" label="Back" class="--back-btn" />
-                            <div class="standard-content--body">
-                                <div class="standard-content--body__form">
-                                    <span v-show="fromCoin === null || (fromCoin.type !== 'new_public_key')" class="lab-input">From</span>
-                                    <q-select v-show="fromCoin === null || (fromCoin.type !== 'new_public_key')" light separator rounded outlined class="select-input" @input="checkGetPairs()" v-model="fromCoin" :options="optionsFrom">
-                                        <template v-slot:option="scope">
-                                            <q-item class="custom-menu" v-bind="scope.itemProps" v-on="scope.itemEvents">
-                                                <q-item-section avatar>
-                                                    <q-icon class="option--avatar option--avatar" :class="scope.opt.value" :name="`img:${scope.opt.image}`" />
-                                                </q-item-section>
-                                                <q-item-section dark>
-                                                    <q-item-label v-html="scope.opt.label" />
-                                                    <q-item-label caption class="ellipsis mw160">{{ scope.opt.value }}</q-item-label>
-                                                </q-item-section>
-                                            </q-item>
-                                        </template>
-                                        <template v-slot:selected>
-                                            <q-item v-if="fromCoin">
-                                                <q-item-section avatar>
-                                                    <q-icon class="option--avatar option--avatar__custom" :name="`img:${fromCoin.image}`" />
-                                                </q-item-section>
-                                                <q-item-section>
-                                                    <q-item-label v-html="fromCoin.label" />
-                                                    <q-item-label caption class="ellipsis mw160">{{ fromCoin.value }}</q-item-label>
-                                                </q-item-section>
-                                            </q-item>
-                                            <q-item v-else>
-                                            </q-item>
-                                        </template>
-                                    </q-select>
-                                    <span v-show="fromCoin !== null && (fromCoin.type === 'new_public_key')" class="lab-input">Your <strong>{{ depositCoin !== null ? depositCoin.value.toUpperCase() : '' }}</strong> return address </span>
-                                    <q-input v-show="fromCoin !== null && (fromCoin.type === 'new_public_key')" v-model="refundAddress.address" @input="verifyAddress()" class="input-input" rounded outlined color="purple" type="text" hint="[ in case the transaction does not complete ]">
-                                        <template v-slot:append>
-                                            <div class="flex justify-end">
-                                                <q-btn color="purple" rounded class="q-mb-sm" @click="fromCoin = null" outlined unelevated flat text-color="black" label="Hide" />
-                                            </div>
-                                        </template>
-                                    </q-input>
-                                    <br v-show="fromCoin !== null && (fromCoin.type === 'new_public_key')">
-                                    <span class="lab-input" v-show="fromCoinMemo" />
-                                    <q-input v-show="fromCoinMemo" class="input-input" rounded outlined color="purple" type="text" v-model="refundAddress.tag" label="Optional tag or memo" hint="some exchanges require this field">
-                                        <template v-slot:append>
-                                            <div class="flex justify-end">
-                                                <q-btn color="purple" rounded class="q-mb-sm" @click="fromCoinMemo = false" outlined unelevated flat text-color="black" label="Hide" />
-                                            </div>
-                                        </template>
-                                    </q-input>
-                                    <br v-show="fromCoinMemo">
-                                    <q-btn v-show="!fromCoinMemo" flat class="q-mt-sm q-mb-sm --next-btn align-left full-width" :icon-right="fromCoinMemo ? 'close':'add'" rounded :label="fromCoinMemo ? 'Hide Tag/Memo':'Add Tag/Memo'" @click="fromCoinMemo = !fromCoinMemo" />
-                                    <br>
-                                    <hr>
-                                    <span v-show="toCoin === null || (toCoin.type !== 'new_public_key')" class="lab-input">To</span>
-                                    <q-select v-show="toCoin === null || (toCoin.type !== 'new_public_key')" light separator rounded outlined class="select-input" v-model="toCoin" @input="updateCoinName()" use-input :options="optionsTo">
-                                        <template v-slot:option="scope">
-                                            <q-item class="custom-menu" v-bind="scope.itemProps" v-on="scope.itemEvents">
-                                                <q-item-section avatar>
-                                                    <q-icon class="option--avatar option--avatar" :class="scope.opt.value" :name="`img:${scope.opt.image}`" />
-                                                </q-item-section>
-                                                <q-item-section dark>
-                                                    <q-item-label v-html="scope.opt.label" />
-                                                    <q-item-label caption class="ellipsis mw160">{{ scope.opt.value }}</q-item-label>
-                                                </q-item-section>
-                                            </q-item>
-                                        </template>
-                                        <template v-slot:selected>
-                                            <q-item v-if="toCoin">
-                                                <q-item-section avatar>
-                                                    <q-icon class="option--avatar option--avatar__custom" :class="toCoin.value" :name="`img:${toCoin.image}`" />
-                                                </q-item-section>
-                                                <q-item-section>
-                                                    <q-item-label v-html="toCoin.label" />
-                                                    <q-item-label caption class="ellipsis mw160">{{ toCoin.value }}</q-item-label>
-                                                </q-item-section>
-                                            </q-item>
-                                            <q-item v-else>
-                                            </q-item>
-                                        </template>
-                                    </q-select>
-                                    <span v-show="toCoin !== null && toCoin.type === 'new_public_key'" class="lab-input">{{ destinationAddressLabel }}</span>
-                                    <q-input v-show="toCoin !== null && toCoin.type === 'new_public_key'" ref="destinationAddressAddress" v-model="destinationAddress.address" :rules="[ val => val.length >= 3 || 'Destination Address Cannot less than 3 characters' ]" @input="verifyAddress()" class="input-input" rounded outlined color="purple" type="text">
-                                        <template v-slot:append>
-                                            <div class="flex justify-end">
-                                                <q-btn color="purple" rounded class="q-mb-sm" @click="toCoin = null" outlined unelevated flat text-color="black" label="Hide" />
-                                            </div>
-                                        </template>
-                                    </q-input>
-                                    <span class="lab-input" v-show="toCoinMemo" />
-                                    <q-input v-show="toCoinMemo" class="input-input" rounded outlined color="purple" type="text" v-model="destinationAddress.tag" label="Optional tag or memo" hint="some exchanges require this field">
-                                        <template v-slot:append>
-                                            <div class="flex justify-end">
-                                                <q-btn color="purple" rounded class="q-mb-sm" @click="toCoinMemo = false" outlined unelevated flat text-color="black" label="Hide" />
-                                            </div>
-                                        </template>
-                                    </q-input>
-                                    <br v-show="toCoinMemo">
-                                    <q-btn v-show="!toCoinMemo" flat class="q-mt-sm q-mb-sm --next-btn align-left full-width" :icon-right="toCoinMemo ? 'close':'add'" rounded :label="toCoinMemo ? 'Hide Tag/Memo':'Add Tag/Memo'" @click="toCoinMemo = !toCoinMemo" />
-                                </div>
-                            </div>
-                            <q-stepper-navigation v-show="true" class="flex justify-end">
-                                <div class="row full-width" style="padding-left: 6px; margin-top: -20px;">
-                                    <div class="q-gutter-sm">
-                                        <q-btn color="white" flat @click="showDisclaimerWrapper = true" class="lower bold" text-color="black" label="Read the disclaimer" />
-                                    </div>
-                                </div>
-                                <div class="row full-width" style="padding-left: 13px; margin-top: 10px;">
-                                    <div class="q-gutter-sm">
-                                        <q-checkbox label="I accept" color="deep-purple-14" v-model="disclaimerCheck" />
-                                    </div>
-                                </div>
-                                <div class="standard-content--footer">
-                                    <q-btn @click="checkAddressMatchCoins()" :disable="!disclaimerCheck" flat class="action-link next" color="black" text-color="white">
-                                        <span class="label">Exchange {{ fromCoinType.toUpperCase() }}
-                                            <q-icon name="keyboard_backspace" color="white" class="left-icon" /> {{ toCoinType.toUpperCase() }}
-                                        </span>
-                                    </q-btn>
-                                </div>
-                            </q-stepper-navigation>
-                        </q-step>
-                        <!-- 5. View Order -->
-                        <q-step default title="View Order" :name="5" prefix="5" :done="step > 5">
-                            <q-btn flat @click="$refs.stepper.previous()" unelevated icon="keyboard_arrow_left" color="primary" label="Back" class="--back-btn" />
-                            <div class="standard-content--body">
-                                <div class="standard-content--body__form" style="margin-left: -35px;">
-                                    <div class="progress-custom-volentix column flex-center">
-                                        <svg class="svg_logo" fill="#7272FA" width="40" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 16 20.58">
-                                            <path d="M199,25.24q0,3.29,0,6.57a.5.5,0,0,1-.18.41l-7.32,6.45a.57.57,0,0,1-.71,0l-7.21-6.1c-.12-.11-.25-.22-.38-.32a.53.53,0,0,1-.22-.47q0-3.83,0-7.66,0-2.69,0-5.39c0-.33.08-.47.29-.51s.33.07.44.37l3.45,8.84c.52,1.33,1,2.65,1.56,4a.21.21,0,0,0,.23.16h4.26a.19.19,0,0,0,.21-.14l3.64-9.7,1.21-3.22c.08-.22.24-.32.42-.29a.34.34,0,0,1,.27.37c0,.41,0,.81,0,1.22Q199,22.53,199,25.24Zm-8.75,12s0,0,0,0,0,0,0,0a.28.28,0,0,0,0-.05l-1.88-4.83c0-.11-.11-.11-.2-.11h-3.69s-.1,0-.13,0l.11.09,4.48,3.8C189.38,36.55,189.8,36.93,190.25,37.27Zm-6.51-16.76h0s0,.07,0,.1q0,5.4,0,10.79c0,.11,0,.16.15.16h4.06c.15,0,.15,0,.1-.16s-.17-.44-.26-.66l-3.1-7.94Zm14.57.06c-.06,0-.06.07-.07.1l-1.89,5q-1.06,2.83-2.13,5.66c-.06.16,0,.19.13.19h3.77c.16,0,.2,0,.2-.2q0-5.3,0-10.59Zm-7.16,17,.05-.11,1.89-5c.05-.13,0-.15-.11-.15h-3.71c-.17,0-.16,0-.11.18.26.65.51,1.31.77,2Zm.87-.3,0,0,5.65-5H194c-.13,0-.16.07-.19.17l-1.59,4.23Zm0,.06h0Z" transform="translate(-183 -18.21)"></path>
-                                        </svg>
-                                        <span class="title">{{ friendlyStatus }}</span>
-                                        <q-linear-progress indeterminate stripe rounded size="md" :value="progress" class="q-mt-md" />
-                                    </div>
-                                    <hr style="height:15px;opacity:0" />
-                                    <div class="text-black">
-                                        <div class="text-h4 --subtitle">
-                                            <ul>
-                                                <li><span>{{exchangeLabel}}</span></li>
-                                            </ul>
-                                        </div>
-                                        <q-input v-model="exchangeAddress.address" readonly rounded class="input-input pr80" outlined color="purple" type="text">
-                                            <template v-slot:append>
-                                                <div class="flex justify-end">
-                                                    <q-btn flat unelevated text-color="grey" @click="copyToClipboard(exchangeAddress.address , 'Exchange Address')" round class="btn-copy" icon="o_file_copy" />
-                                                </div>
-                                            </template>
-                                        </q-input>
-                                    </div>
-                                </div>
-                            </div>
-                        </q-step>
-                    </q-stepper>
-                </div>
-                <br><br><br>
-            </div>
-        </div>
-    </div>
-    <q-dialog persistent v-model="showDisclaimerWrapper">
-        <q-card class="q-pa-md" style="width: 700px; max-width: 92vw;">
-            <q-toolbar>
-                <q-avatar><img src="statics/icon.png"></q-avatar>
-                <q-toolbar-title><span class="text-weight-bold">Disclaimer</span> </q-toolbar-title>
-            </q-toolbar>
-            <q-card-section class="text-h6">
-                <div class="parg">
-                    This transaction is carried out using an open API linked to various Exchanges.<br>
-                    By sending coins to the above address, you agreed to <a href="https://coinswitch.co/terms" target="_blank">the terms and conditions</a> of the selected provider.
-                </div>
-            </q-card-section>
-            <q-card-actions align="right" class="q-pr-sm">
-                <q-btn flat label="Accept" class="accept-disclaimer" @click="disclaimerCheck = true" color="primary" v-close-popup />
-            </q-card-actions>
-        </q-card>
-    </q-dialog>
-</div>
-</template>
-
-<script>
-import store from '@/store'
-// import { userError } from '@/util/errorHandler'
-
-const url = 'https://api.coinswitch.co'
-let headers = {
-  'x-api-key': process.env[store.state.settings.network].COINSWITCH_APIKEY
-}
-
-const typeUpper = function (thing) {
-  if (typeof thing === 'string' && thing.length >= 1) {
-    return thing.toUpperCase()
-  } else {
-    return ''
-  }
-}
-import DexInteraction from '../../mixins/DexInteraction'
-import Lib from '@/util/walletlib'
-import EosWrapper from '@/util/EosWrapper'
-const eos = new EosWrapper()
-export default {
-  components: {
-
-  },
-  data () {
-    return {
-      pStep: 1,
-      fetchingRate: false,
-      model: null,
-      receiveCoin: 'ETH',
-      payCoin: 'BTC',
-      osName: '',
-      disclaimerCheck: false,
-      showDisclaimerWrapper: false,
-      fromCoin: null,
-      fromCoinAmount: 0,
-      toCoinAmount: 0,
-      fromCoinType: 'BTC',
-      toCoinType: 'ETH',
-      toCoin: null,
-      progress: 0.2,
-      to: '',
-      amount: '',
-      memo: '',
-      optionsFrom: [],
-      optionsTo: [],
-      minimizedModal: false,
-      params: null,
-      tableData: [],
-      currentAccount: null,
-      goBack: '/verto/dashboard',
-      fetchCurrentWalletFromState: true,
-      fromCoinMemo: false,
-      toCoinMemo: false,
-      // -------------------------
-      timeoutNotif: 0,
-      notif: false,
-      step: 1,
-      optionsSanitize: false,
-      spinnervisible: false,
-      lastChangedValue: 'deposit',
-      coins: [],
-      depositCoin: {
-        'label': 'Bitcoin',
-        'value': 'btc',
-        'image': 'https://files.coinswitch.co/public/coins/btc.png'
-      },
-      depositQuantity: 0,
-      depositCoinOptions: null,
-      depositCoinUnfilter: null,
-      destinationCoin: {
-        'label': 'Ethereum',
-        'value': 'eth',
-        'image': 'https://files.coinswitch.co/public/coins/eth.png'
-      },
-      destinationQuantity: 0,
-      destinationCoinOptions: null,
-      destinationCoinUnfilter: null,
-      rateData: null,
-      rateDataTemplate: {
-        rate: 1,
-        minerFee: 0,
-        limitMinDepositCoin: 0,
-        limitMaxDepositCoin: 1,
-        limitMinDestinationCoin: 1,
-        limitMaxDestinationCoin: 2
-      },
-      destinationAddress: {
-        address: '',
-        tag: null
-      },
-      refundAddress: {
-        address: '',
-        tag: null
-      },
-      exchangeAddress: {
-        address: '',
-        tag: null
-      },
-      expectedDepositCoinAmount: 0,
-      expectedDestinationCoinAmount: 0,
-      orderId: null,
-      status: null,
-      requestStop: false,
-      destinationCoinAmount: null,
-      vtxEosPrice: null,
-      rateDataVtx: null,
-      rateDataEos: null
-    }
-  },
-  updated () {},
-  watch: {
-    depositCoin (val) {
-      console.log('new val', val)
-      this.payCoin = val.value.toUpperCase()
-    },
-    destinationCoin (val) {
-      console.log('new val', val)
-      this.receiveCoin = val.value.toUpperCase()
-    },
-    fromCoin (val) {
-      if (val.type !== 'new_public_key') {
-        this.fromCoinType = this.fromCoin.type
-      } else {
-        return 'EOS'
-      }
-    },
-    toCoin (val) {
-      if (val.type !== 'new_public_key') {
-        this.toCoinType = this.toCoin.type
-      } else {
-        return 'VTX'
-      }
-    }
-  },
-  async created () {
-    this.getWindowWidth()
-    window.addEventListener('resize', this.getWindowWidth)
-    this.params = this.$store.state.currentwallet.params
-    this.tableData = await this.$store.state.wallets.tokens
-    let self = this
-    this.tableData.map(token => {
-      self.optionsFrom.push({
-        label: token.name.toLowerCase(),
-        value: token.key,
-        image: token.icon,
-        type: token.type
-      })
-      self.optionsTo.push({
-        label: token.name.toLowerCase(),
-        value: token.key,
-        privateKey: token.privateKey,
-        image: token.icon,
-        type: token.type
-      })
-    })
-    this.optionsFrom.unshift({
-      label: 'Other public key / Account',
-      value: '',
-      image: '/statics/img/door-key.png',
-      type: 'new_public_key'
-    })
-    this.optionsTo.unshift({
-      label: 'Other public key / Account',
-      value: '',
-      image: '/statics/img/door-key.png',
-      type: 'new_public_key'
-    })
-    this.currentAccount = this.tableData.find(w => w.chain === this.params.chainID && w.type === this.params.tokenID && (
-      w.chain === 'eos' ? w.name.toLowerCase() === this.params.accountName : w.key === this.params.accountName))
-    // console.log('this.currentAccount', this.currentAccount)
-    if (this.currentAccount !== null && this.currentAccount !== undefined) {
-      // this.fromCoin = {
-      //   label: this.currentAccount.name,
-      //   value: this.currentAccount.key,
-      //   image: this.currentAccount.icon,
-      //   type: this.currentAccount.type
-      // }
-      this.goBack = this.fetchCurrentWalletFromState ? `/verto/wallets/${this.params.chainID}/${this.params.tokenID}/${this.params.accountName}` : '/verto/dashboard'
-    }
-    this.$q.notify.registerType('my-notif', {
-      icon: 'announcement',
-      progress: false,
-      persistent: true,
-      timeout: this.timeoutNotif,
-      color: 'deep-purple-14',
-      textColor: 'white',
-      position: 'bottom-right'
-    })
-    this.checkGetPairs()
-    this.checkToGetRate()
-  },
-  computed: {
-    getStatus () {
-      let value = 0
-
-      switch (this.status) {
-        case null:
-        case 'no_deposit':
-        case 'failed':
-        case 'refunded':
-        case 'timeout':
-          value = 0
-          break
-        case 'confirming':
-          value = 25
-          break
-        case 'exchanging':
-          value = 50
-          break
-        case 'sending':
-          value = 75
-          break
-        case 'complete':
-          value = 100
-          break
-      }
-
-      return value
-    },
-    friendlyStatus () {
-      let value = ''
-
-      switch (this.status) {
-        case null:
-          value = ''
-          break
-        case 'limit_exceeded':
-          value = 'The destination amount limit was exceeded'
-          break
-        case 'invalid_address':
-          value = 'The destination address is invalid'
-          break
-        case 'no_deposit':
-          value = 'No deposit detected yet'
-          break
-        case 'failed':
-          value = 'The transaction has failed'
-          break
-        case 'refunded':
-          value = 'The transaction has been refunded'
-          break
-        case 'timeout':
-          value = 'The transaction has timed out'
-          break
-        case 'confirming':
-          value = 'The transaction is confirming'
-          break
-        case 'exchanging':
-          value = 'The transaction is exchanging'
-          break
-        case 'sending':
-          value = 'The coins are being sent'
-          break
-        case 'complete':
-          value = 'The transaction is complete'
-          break
-      }
-
-      return value
-    },
-    trackColor () {
-      let value = ''
-
-      switch (this.status) {
-        case null:
-        case 'no_deposit':
-        case 'confirming':
-        case 'exchanging':
-        case 'limit_exceeded':
-        case 'invalid_address':
-        case 'sending':
-        case 'complete':
-          value = 'white'
-          break
-        case 'failed':
-        case 'refunded':
-        case 'timeout':
-          value = 'red'
-          break
-      }
-
-      return value
-    },
-    logoUrl () {
-      if (this.destinationCoin != null) {
-        return this.coins.filter(coins => coins.symbol === this.destinationCoin.value)[0].logoUrl
-      } else {
-        return '/static/icon.png'
-      }
-    },
-    exchangeLabel () {
-      if (this.depositCoin != null) {
-        return 'Complete this exchange by sending ' + this.expectedDepositCoinAmount + ' ' + typeUpper(this.depositCoin.value) + ' to this address within the next 1 hour'
-      } else {
-        return 'Complete this exchange by sending the coins to this address within the next 1 hour'
-      }
-    },
-    depositQuantityLabel () {
-      if (this.depositCoin != null) {
-        return typeUpper(this.depositCoin.value) + ' to Send'
-      } else {
-        return 'Coin to Send'
-      }
-    },
-    destinationQuantityLabel () {
-      if (this.destinationCoin != null) {
-        return typeUpper(this.destinationCoin.value) + ' to Receive'
-      } else {
-        return 'Coin to Receive'
-      }
-    },
-    returnAddressLabel () {
-      if (this.depositCoin != null) {
-        return 'Your ' + typeUpper(this.depositCoin.value) + ' return address [in case the transaction does not complete]'
-      } else {
-        return 'Your return address [in case the transaction does not complete]'
-      }
-    },
-    destinationAddressLabel () {
-      if (this.destinationCoin != null) {
-        return 'Address to receive new ' + typeUpper(this.destinationCoin.value)
-      } else {
-        return 'Address to receive new coin'
-      }
-    }
-  },
-  async mounted () {
-    const self = this
-
-    self.coins = this.getAllCoins()
-    self.depositCoinOptions = self.coins
-    self.destinationCoinUnfilter = self.coins
-    self.depositCoinUnfilter = self.depositCoinOptions
-
-    if (this.$store.state.settings.dexData.depositCoin) {
-      // console.log(this.$route.params.depositCoin, this.$route.params.destinationCoin)
-      this.depositCoin = this.$store.state.settings.coins.coinswitch.find(o => o.value.toLowerCase() === this.$store.state.settings.dexData.depositCoin.value.toLowerCase())
-      this.checkGetPairs()
-      this.checkToGetPairs()
-      this.step = 2
-      if (this.$store.state.settings.dexData.destinationCoin) {
-        this.destinationCoin = this.$store.state.settings.coins.coinswitch.find(o => o.value.toLowerCase() === this.$store.state.settings.dexData.destinationCoin.value.toLowerCase())
-        this.updateCoinName()
-        this.checkToGetRate()
-        this.step = 3
-      }
-    } else {
-      this.depositCoin = this.depositCoinOptions[0]
-      this.destinationCoin = this.depositCoinOptions[1]
-    }
-  },
-  methods: {
-    getWindowWidth () {
-      this.screenSize = document.querySelector('#q-app').offsetWidth
-    },
-    copyToClipboard (key, copied) {
-      this.$clipboardWrite(key)
-      this.$q.notify({
-        message: copied ? copied + ' Copied' : 'Key Copied',
-        timeout: 2000,
-        icon: 'check',
-        textColor: 'white',
-        type: 'warning',
-        position: 'top'
-      })
-    },
-    checkAddressMatchCoins () {
-      this.pStep = 3
-      this.postOrder()
-      // this.$refs.stepper.next()
-    },
-    checkDepositAndDestination () {
-      if (this.destinationAddress.address === '' || this.depositCoin.address === '') {}
-    },
-    updateCoinName () {
-      if (this.destinationCoin !== null) {
-        this.toCoinType = this.destinationCoin.value
-        this.getPairs()
-        this.checkToGetRate()
-      } else {
-        this.toCoinType = this.toCoin.type
-      }
-    },
-    checkGetPairs () {
-      if (this.depositCoin !== null) {
-        this.fromCoinType = this.depositCoin.value
-        this.getPairs()
-        this.checkToGetPairs()
-        this.checkToGetRate()
-      } else {
-        this.fromCoinType = this.fromCoin.type
-      }
-    },
-    switchAmounts () {
-      let depositCoinVar = this.depositCoin
-      this.depositCoin = this.destinationCoin
-      this.destinationCoin = depositCoinVar
-
-      let fromCoinTypeVar = this.fromCoinType
-      this.fromCoinType = this.toCoinType
-      this.toCoinType = fromCoinTypeVar
-
-      let depositQuantityVar = this.depositQuantity
-      this.depositQuantity = this.destinationQuantity
-      this.destinationQuantity = depositQuantityVar
-
-      this.getRate()
-    },
-    filterDepositCoin (val, update, abort) {
-      update(() => {
-        const needle = val.toLowerCase()
-        this.depositCoinOptions = this.depositCoinUnfilter.filter(v => v.value.toLowerCase().indexOf(needle) > -1)
-      })
-    },
-    filterDestinationCoin (val, update, abort) {
-      update(() => {
-        const needle = val.toLowerCase()
-        this.destinationCoinOptions = this.destinationCoinUnfilter.filter(v => v.value.toLowerCase().indexOf(needle) > -1)
-      })
-    },
-    copy2clip (value) {
-      // more generic copy
-      this.$clipboardWrite(value)
-      this.$q.notify({
-        message: this.$t('Main.copied'),
-        color: 'positive'
-      })
-    },
-    checkToPostOrder () {
-      if (this.$refs.depositQuantity.hasError || this.$refs.destinationQuantity.hasError) {
-        // userError()
-        // console.error('There is a problem with the quantities')
-      } else {
-        this.postOrder()
-        // this.$refs.stepper.next()
-      }
-    },
-    checkToGetPairs () {
-      if (this.depositCoin === null) {
-        // userError('There is a problem with the coin selection')
-        // console.error('There is a problem with the coin selection')
-      } else {
-        this.getPairs()
-        // this.$refs.stepper.next()
-      }
-    },
-    checkToGetRate () {
-      // if (this.$refs.destinationAddressAddress.hasError || this.destinationAddress.address === '' ||
-      // console.log('this.depositCoin.value', this.depositCoin.value)
-      let self = this
-      this.optionsFrom = []
-      this.optionsTo = []
-      this.tableData.map(token => {
-        if (this.depositCoin.value.toLowerCase() === token.type) {
-          self.optionsFrom.push({
-            label: token.name.toLowerCase(),
-            value: token.chain === 'eos' ? token.name.toLowerCase() : token.key,
-            key: token.key,
-            image: token.icon,
-            type: token.type
-          })
-        }
-        if ((this.destinationCoin.value.toLowerCase() === token.type) || (this.destinationCoin.value.toLowerCase() === 'vtx' && token.type === 'eos')) {
-          self.optionsTo.push({
-            label: token.name.toLowerCase(),
-            value: token.chain === 'eos' ? token.name.toLowerCase() : token.key,
-            key: token.key,
-            privateKey: token.privateKey,
-            image: token.icon,
-            type: token.type
-          })
-        }
-      })
-      this.optionsFrom.unshift({
-        label: 'Other public key / Account',
-        value: '',
-        image: '/statics/img/door-key.png',
-        type: 'new_public_key'
-      })
-      this.optionsTo.unshift({
-        label: 'Other public key / Account',
-        value: '',
-        image: '/statics/img/door-key.png',
-        type: 'new_public_key'
-      })
-      if (this.destinationCoin === null) {
-        // userError()
-        // console.error('There is a problem with the destination address or the coin is not selected')
-      } else {
-        this.getRate()
-        // this.$refs.stepper.next()
-      }
-    },
-    verifyAddress () {
-      // check validity of all keys
-    },
-    quantityFromDeposit () {
-      // deal with precision
-      this.destinationQuantity = (+this.depositQuantity * +this.rateData.rate) - +this.rateData.minerFee
-      this.lastChangedValue = 'deposit'
-    },
-    quantityFromDestination () {
-      // deal with precision
-      this.depositQuantity = (+this.destinationQuantity + +this.rateData.minerFee) / +this.rateData.rate
-      this.lastChangedValue = 'destination'
-    },
-    orderStatus () {
-      const self = this
-      this.$axios.get(url + '/v2/order/' + this.orderId, {
-        headers
-      }).then(function (result) {
-        self.status = result.data.data.status
-        if (self.status === 'no_deposit') {
-          if (!self.notif) {
-            self.triggerCustomRegisteredType1()
-          }
-        } else {
-          self.timeoutNotif = 1000
-        }
-        if (self.status === 'no_deposit' ||
-                    self.status === 'confirming' ||
-                    self.status === 'exchanging' ||
-                    self.status === 'sending') {
-          setTimeout(() => {
-            self.orderStatus()
-          }, 30000)
-        }
-        if (self.status === 'complete' && self.destinationCoin.value === 'vtx') {
-          self.destinationCoinAmount = Math.trunc(result.data.data.destinationCoinAmount * 10000) / 10000
-          self.orderVTX()
-        }
-      })
-    },
-    async orderVTX () {
-      // check balance then...
-      // let eosBal = Lib.balance('eos', this.toCoin.value, 'eos')
-      let eosBal = (await eos.getCurrencyBalanceP(this.toCoin.value)).toString().split(' ')[0]
-      console.log('eosBal', eosBal)
-      if (+eosBal < +this.destinationCoinAmount) {
-        console.log('eos balance is yet to low to proceed: ', eosBal)
-        setTimeout(() => {
-          self.orderVTX()
-        }, 1000)
-      } else {
-        Lib.send(
-          'eos',
-          'eos',
-          this.toCoin.value,
-          'swap.defi', // 'newdexpublic',
-          this.destinationCoinAmount,
-          'swap,0,448', // '{"type":"buy-market","symbol":"volentixgsys-vtx-eos","price":"0.00000","channel":"dapp","ref":"verto"}',
-          this.toCoin.privateKey,
-          'eosio.token'
-        ).then(result => {
-          console.log('send result', result)
-          if (result.success) {
-            this.$q.notify({
-              message: 'Your VTX have been received',
-              color: 'positive'
-            })
-          } else {
-            this.$q.notify({
-              message: 'Could not convert EOS to VTX',
-              color: 'negative',
-              type: 'warning'
-            })
-          }
-        })
-      }
-    },
-    postOrder () {
-      const self = this
-      console.log('postOrder called', self.depositQuantity, ' + ', self.destinationQuantity)
-      let depositCoinAmount = null
-      let destinationCoinAmount = null
-
-      if (self.lastChangedValue === 'deposit') {
-        depositCoinAmount = self.depositQuantity
-      } else {
-        destinationCoinAmount = self.destinationCoin.value === 'vtx' ? self.destinationQuantity * self.vtxEosPrice : self.destinationQuantity
-      }
-
-      this.refundAddress.address = this.refundAddress.address === '' ? this.fromCoin.value : this.refundAddress.address
-      // console.log('this.refundAddress', this.refundAddress)
-      this.destinationAddress.address = this.destinationAddress.address === '' ? this.toCoin.value : this.destinationAddress.address
-
-      this.$axios.post(url + '/v2/order', {
-        depositCoin: self.depositCoin.value,
-        destinationCoin: self.destinationCoin.value === 'vtx' ? 'eos' : self.destinationCoin.value,
-        depositCoinAmount,
-        destinationCoinAmount,
-        destinationAddress: self.destinationAddress,
-        refundAddress: self.refundAddress
-      }, {
-        headers
-      })
-        .then((response) => {
-          // console.log('response - order', response)
-          self.orderId = response.data.data.orderId
-          self.exchangeAddress = response.data.data.exchangeAddress
-          self.expectedDepositCoinAmount = response.data.data.expectedDepositCoinAmount
-          self.expectedDestinationCoinAmount = response.data.data.expectedDestinationCoinAmount
-
-          this.orderStatus()
-        })
-        .catch((err) => {
-          if (err) {}
-          // userError()
-          // console.error('There was a problem posting the order', err)
-        })
-    },
-    getPairs () {
-      const self = this
-      this.$axios.post(url + '/v2/pairs', {
-        depositCoin: self.depositCoin.value
-      }, {
-        headers
-      })
-        .then((response) => {
-          // console.log('------------Response------------', response)
-          let inject = {}
-          self.destinationCoinOptions = response.data.data.map(function (coin) {
-            if (coin.isActive === true) {
-              let row = {
-                'label': self.coins.filter(coins => coins.symbol === coin.destinationCoin)[0].name,
-                'value': coin.destinationCoin,
-                'image': self.coins.filter(coins => coins.symbol === coin.destinationCoin)[0].logoUrl
-              }
-              if (coin.destinationCoin === 'eos') {
-                inject = {
-                  'label': 'Volentix',
-                  'value': 'vtx',
-                  'chain': 'eos',
-                  'image': '/statics/vtx_icon.png'
-                }
-              }
-              return row
-            } // deal with false, should not create empty option.
-          }).filter(function (el) {
-            return el != null
-          }).sort(function (a, b) {
-            if (a.label.toLowerCase() < b.label.toLowerCase()) {
-              return -1
-            }
-            return 1
-          })
-          self.destinationCoinOptions.push(inject)
-          self.destinationCoinOptions.sort(function (a, b) {
-            if (a.label.toLowerCase() < b.label.toLowerCase()) {
-              return -1
-            }
-            return 1
-          })
-          console.log(self.destinationCoinOptions, 'self.destinationCoinOptions')
-          self.destinationCoinUnfilter = self.destinationCoinOptions
-        })
-        .catch((err) => {
-          if (err) {}
-          // userError()
-          // console.error('There was a problem getting the destination coins', err)
-        })
-    },
-    async getRate () {
-      const self = this
-      if (self.destinationCoin.value === 'vtx') {
-        this.vtxEosPrice = (await this.$axios.get(process.env[this.$store.state.settings.network].CACHE + 'https://api.newdex.io/v1/price?symbol=volentixgsys-vtx-eos')).data.data.price
-      }
-      this.$axios.post(url + '/v2/rate', {
-        depositCoin: self.depositCoin.value,
-        destinationCoin: self.destinationCoin.value === 'vtx' ? 'eos' : self.destinationCoin.value
-      }, {
-        headers
-      })
-        .then((response) => {
-          self.rateData = response.data.data
-          if (self.destinationCoin.value === 'vtx') {
-            self.rateDataVtx = {
-              limitMaxDepositCoin: self.rateData.limitMaxDepositCoin,
-              limitMaxDestinationCoin: self.rateData.limitMaxDestinationCoin / self.vtxEosPrice,
-              limitMinDepositCoin: self.rateData.limitMinDepositCoin,
-              limitMinDestinationCoin: self.rateData.limitMinDestinationCoin / self.vtxEosPrice,
-              minerFee: self.rateData.minerFee,
-              rate: self.rateData.rate / self.vtxEosPrice
-            }
-            self.rateDataEos = self.rateData
-            self.rateData = self.rateDataVtx
-          }
-          console.log('self.rateData -------------- ', self.rateData)
-        })
-        .catch((err) => {
-          if (err) {}
-          // userError()
-          // console.error('There was a problem getting the rate data', err)
-        })
-    },
-    triggerCustomRegisteredType1 () {
-      this.notif = true
-      this.$q.notify({
-        type: 'my-notif',
-        closeBtn: '+',
-        classes: 'exchange-notif',
-        message: `This may take a few minutes.`,
-        caption: 'If you leave this page, you will no longer be able to track the status of this transaction.'
-      })
-    },
-    addSearchField () {
-      console.log('addSearchField triggered')
-    },
-    changeText () {
-      console.log('changeText triggered')
-    },
-    triggerPayCoinSelect () {
-      // console.log('triggerPayCoinSelect triggered')
-      document.querySelector('.pay-coin-select .q-field__control').dispatchEvent(new Event('click'))
-      setTimeout(() => {
-        let payCoinSelectPopup = document.querySelector('.pay-coin-select-popup')
-        let paycoinSearch = document.querySelector('.paycoin-search').cloneNode(true)
-        // console.log('paycoinSearch', paycoinSearch)
-        payCoinSelectPopup.prepend(paycoinSearch)
-        payCoinSelectPopup.querySelector('.paycoin-search').classList.remove('hidden')
-      }, 100)
-      // paycoin-search
-      // pay-coin-select-popup
-    },
-    triggerReceiveCoinSelect () {
-      // console.log('triggerReceiveCoinSelect triggered')
-      document.querySelector('.receive-coin-select .q-field__control').dispatchEvent(new Event('click'))
-      setTimeout(() => {
-        let receiveCoinSelectPopup = document.querySelector('.receive-coin-select-popup')
-        let receivecoinSearch = document.querySelector('.receivecoin-search').cloneNode(true)
-        // console.log('receivecoinSearch', receivecoinSearch)
-        receiveCoinSelectPopup.prepend(receivecoinSearch)
-        receiveCoinSelectPopup.querySelector('.receivecoin-search').classList.remove('hidden')
-      }, 100)
-      // paycoin-search
-      // pay-coin-select-popup
-    }
-  },
-  mixins: [DexInteraction]
-}
-</script>
-
-<style lang="scss" scoped>
-@import "~@/assets/styles/variables.scss";
-
-/deep/ .wallets-wrapper {
-    padding-bottom: 0px !important;
-}
-
-/deep/ .wallets-wrapper--list {
-    box-shadow: none;
-    margin-top: 0px;
-}
-
-.marg {
-    /deep/ .profile-wrapper {
-        &--header {
-            margin-bottom: 0px;
-        }
-    }
-}
-
-.mobile-pad {
-    padding-bottom: 100px;
-    background: #f3f3f3 !important
-}
-
-.desktop-version {
-    background: #E7E8E8;
-    padding-top: 13vh;
-    padding-left: 12vh;
-    padding-bottom: 50px;
-    padding-right: 2%;
-}
-
-.desktop-card-style {
-    height: 100%;
-}
-
-.standard-content {
-    padding: 5% 10%;
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-
-    // justify-content: space-start;
-    // min-height: calc(100vh + 100px) !important;
-    // padding-bottom: 100px;
-    @media screen and (min-width: 768px) {
-        padding: 2%;
-        flex-direction: column;
-        justify-content: flex-start;
-        min-height: unset !important;
-        padding-bottom: 20px;
-    }
-
-    .exchange_picto {
-        margin-top: -60px;
-
-        img {
-            width: 100%;
-            max-width: 270px;
-            margin-top: 20px;
-            margin-bottom: -30px;
-        }
-
-        @media screen and (min-width: 768px) {
-            margin-top: -20px;
-
-            img {
-                width: 100%;
-                max-width: 380px;
-            }
-        }
-    }
-
-    &--title {
-        font-size: 35px;
-        font-weight: $bold;
-        position: relative;
-        line-height: 50px;
-        font-family: $Titillium;
-        margin-top: 0px;
-        margin-bottom: 0px;
-
-        @media screen and (min-width: 768px) {
-            margin-top: -20px;
-            font-size: 25px;
-        }
-
-        .btn-align-left {
-            position: absolute;
-            left: -15px;
-            top: 10px;
-        }
-    }
-
-    &--desc {
-        margin-top: -20px;
-        margin-bottom: 40px;
-        font-size: 18px;
-        font-weight: $regular;
-        position: relative;
-        line-height: 26px;
-        font-family: $Titillium;
-        color: $mainColor;
-    }
-
-    &--body {
-        .progress-custom-volentix {
-            max-width: 150px;
-            margin: auto;
-            margin-top: 20px;
-
-            .title {
-                font-size: 13px;
-                color: #7272FA;
-                font-weight: $light;
-                font-family: $Titillium;
-                margin-bottom: -10px;
-                margin-top: 10px;
-            }
-
-            /deep/ .q-linear-progress__model {
-                background: #7272FA;
-            }
-        }
-
-        &__img {
-            min-height: 250px;
-
-            img {
-                max-width: 90%;
-            }
-        }
-
-        .rate-value {
-            padding: 0px 0px;
-            font-family: $Titillium;
-            font-weight: $regular;
-            color: #B0B0B0;
-            font-size: 16px;
-            margin-bottom: -20px;
-            margin-top: -20px;
-        }
-
-        &__form {
-            .pay-get-wrapper {
-                // border-bottom: 1px solid rgba(#707070, .4);
-                padding: 0px;
-                padding-bottom: 10px;
-                margin-bottom: 20px;
-
-                .pay-wrapper {}
-
-                .exchange-btn {
-                    height: fit-content;
-                    align-self: flex-end;
-                    padding-top: 15px;
-                    transform: rotate(90deg);
-                    margin: auto;
-                    margin-top: 10px;
-                    padding: 12px 10px 0px 10px;
-
-                    .left-icon {
-                        transform: scaleX(-1);
-                        margin-left: 0px;
-                        margin-right: -26px;
-                        margin-top: -24px;
-                        color: #7272FA;
-                    }
-
-                    .right-icon {
-                        color: #7272FA;
-                    }
-                }
-
-                .label {
-                    font-size: 17px;
-                    color: #2A2A2A;
-                    font-family: $Titillium;
-                    font-weight: $light;
-                }
-
-                .value {
-                    font-size: 33px;
-                    color: #B0B0B0;
-                    font-family: $Titillium;
-                    font-weight: $regular;
-                }
-
-                .get-wrapper {}
-            }
-
-            /deep/ .q-field__native {
-                padding-left: 8px;
-                font-size: 16px;
-                font-weight: $regular;
-            }
-
-            /deep/ .q-field__label {
-                font-family: $Titillium;
-                font-weight: $regular;
-                font-size: 18px;
-                padding-left: 10px;
-                margin-top: -2px;
-            }
-
-            .input-input {
-                height: 50px;
-                padding-bottom: 0px;
-
-                /deep/ .q-field__control {
-                    height: 50px;
-                    min-height: unset;
-
-                    @media screen and (min-width: 1024px) {
-                        margin-left: 20px;
-                    }
-                }
-
-                .btn-copy {
-                    // height: 30px;
-                    position: relative;
-                    top: -3px;
-                    margin-right: 0px;
-                    padding: 6px 13px;
-                }
-
-                .qr-btn {
-                    width: 30px;
-                    height: 30px;
-                    padding-right: 8px;
-                    display: flex;
-                    flex-direction: column;
-                    justify-content: center;
-                    align-items: center;
-
-                    img {
-                        width: 25px;
-                        height: 25px;
-                        position: relative;
-                        right: -5px;
-                        transform: scale3d(1, 1, 1);
-                    }
-                }
-            }
-
-            /deep/ .option--avatar {
-                border: 1px solid;
-                width: 35px;
-                height: 35px;
-                max-width: 40px;
-                padding: 0px;
-                min-width: unset;
-                display: flex;
-                flex-direction: column;
-                justify-content: center;
-                align-items: center;
-                border-radius: 50px;
-                overflow: hidden;
-                margin-left: -10px;
-                padding-right: 0px;
-
-                &.vtx {
-                    background-color: #000;
-                    padding: 3px;
-                }
-            }
-        }
-    }
-
-    &--footer {
-        display: flex;
-        flex-direction: row;
-        justify-content: center;
-        align-items: center;
-        min-height: 100px;
-
-        .action-link {
-            height: 54px;
-            text-transform: initial !important;
-            font-size: 16px;
-            letter-spacing: .5px;
-            border-radius: 40px;
-            width: fit-content;
-            margin-left: 0px;
-            padding-left: 20px;
-            padding-right: 20px;
-            background-color: #7272FA !important;
-
-            .left-icon {
-                transform: scaleX(-1);
-                margin-left: 6px;
-                margin-right: 6px;
-                margin-top: -1px;
-            }
-
-            // &.next{
-            //   background-color: #7900FF !important;
-            // }
-            // &.back{
-            //   background-color: #B0B0B0 !important;
-            // }
-        }
-
-    }
-}
-
-.select-input {
-    border-radius: 100px !important;
-    $height: 50px;
-    height: $height;
-
-    /deep/ .q-field__marginal {
-        height: $height;
-        min-height: unset;
-    }
-
-    /deep/ .q-field__control {
-        height: $height;
-        min-height: unset;
-
-        .q-field__native {
-            padding-left: 0px;
-            padding-top: 0px;
-            padding-bottom: 0px;
-            height: $height;
-            min-height: unset;
-
-            .q-item {
-                padding: 0px;
-                padding-left: 18px;
-                min-height: $height;
-                padding-bottom: 0px;
-
-                .q-item__section {
-                    padding-right: 0px;
-                    min-width: 36px;
-
-                    .q-item__label+.q-item__label {
-                        margin-top: 0px;
-                    }
-                }
-            }
-        }
-    }
-}
-
-.lab-input {
-    font-family: $Titillium;
-    font-weight: $regular;
-    font-size: 16px;
-    color: black;
-    padding-left: 20px;
-    padding-bottom: 5px;
-    display: block;
-    margin-top: 10px;
-    line-height: 17px;
-}
-
-.tlab-input {
-    font-family: $Titillium;
-    font-weight: $bold;
-    font-size: 26px;
-    color: black;
-    padding-left: 20px;
-    padding-bottom: 25px;
-    display: block;
-    margin-top: 0px;
-    line-height: 17px;
-}
-
-.sublab-input {
-    font-family: $Titillium;
-    font-weight: $regular;
-    font-size: 14px;
-    color: #333;
-    padding-left: 20px;
-    padding-bottom: 5px;
-    display: block;
-    margin-top: -50px;
-    line-height: 17px;
-}
-
-.mw200 {
-    max-width: 220px;
-}
-
-.mw160 {
-    max-width: 160px;
-}
-
-.mw100 {
-    max-width: 110px;
-}
-
-.pl0 {
-    padding-left: 0px !important;
-
-    /deep/ .q-field__native {
-        padding-left: 1px !important;
-    }
-}
-
-.align-left {
-    /deep/ .q-btn__content {
-        text-align: left;
-        justify-content: start;
-        padding-left: 7px;
-
-        .q-icon {
-            font-size: 1em;
-        }
-    }
-}
-
-.--next-btn {
-    text-transform: initial !important;
-}
-
-.chain-tools-wrapper {
-
-    // padding: 0px 6%;
-    &--list {
-        &__hide-chain-tools {
-            text-transform: initial !important;
-            margin-top: 0px;
-            margin-bottom: 10px;
-            color: #7272FA !important;
-        }
-
-        .list-wrapper {
-            overflow: hidden;
-            visibility: hidden;
-            height: 0px;
-            opacity: 0;
-            transform: translateY(-20px) scaleY(.5);
-            transform-origin: top;
-            transition: ease transform .3s, ease opacity .4s;
-
-            @media screen and (min-width: 1024px) {
-                overflow: visible;
-            }
-
-            &--chain {
-                &__type {
-                    background-color: #fff;
-                    margin-bottom: 10px;
-                    border-radius: 0px 0px 10px 10px;
-                    padding: 3% 8%;
-                    box-shadow: 0px 4px 16px 0px rgba(black, .09);
-                    font-family: $Titillium;
-                    font-size: 20px;
-                    color: #2A2A2A;
-                    font-weight: $bold;
-
-                    b {
-                        color: #7272FA;
-                        text-transform: uppercase;
-                    }
-
-                    .chain {}
-
-                    .token {}
-                }
-
-                &__coming-soon {
-                    ul {
-                        list-style: none;
-                        padding: 0px;
-                        margin: 0px;
-                        padding: 5% 6%;
-
-                        li {
-                            &:not(:last-child) {
-                                border-bottom: 1px solid #707070;
-                            }
-
-                            .btn-soon {
-                                text-transform: initial !important;
-                                padding: 20px 10px;
-                                border-radius: 0px;
-
-                                /deep/ .q-btn__content {
-                                    display: flex;
-                                    justify-content: space-between;
-                                    flex-direction: row;
-                                    align-items: center;
-                                    align-content: center;
-                                }
-
-                                .title {
-                                    font-size: 20px;
-                                    color: #454F63;
-                                    font-weight: $bold;
-                                    margin-right: auto;
-                                }
-
-                                .soon {
-                                    font-size: 16px;
-                                    color: #B0B0B0;
-                                    display: flex;
-                                    justify-content: space-between;
-                                    flex-direction: row;
-                                    align-items: center;
-                                    align-content: center;
-                                }
-
-                                .icon {
-                                    color: #78849E;
-                                    position: relative;
-                                    top: 2px;
-                                    margin-left: 20px;
-                                }
-                            }
-                        }
-                    }
-                }
-
-                &__eos-to-vtx-convertor {
-                    background-color: #fff;
-                    margin-bottom: 10px;
-                    border-radius: 10px;
-                    padding: 1% 2%;
-                    box-shadow: 0px 4px 16px 0px rgba(black, .09);
-
-                    @media screen and (min-width: 1024px) {
-                        box-shadow: none;
-                    }
-
-                    &--title {
-                        font-size: 22px;
-                        font-family: $Titillium;
-                        font-weight: $bold;
-                        color: #2A2A2A;
-                        margin: 0px;
-                        padding-left: 22px;
-                        margin-top: 3px;
-                        position: relative;
-                    }
-
-                    /deep/ .q-stepper__step {
-                        position: relative;
-                    }
-
-                    /deep/ .q-stepper--vertical .q-stepper__dot:before,
-                    /deep/ .q-stepper--vertical .q-stepper__dot:after {
-                        content: '';
-                        transform: translateX(-50%);
-                        width: 18px;
-                        background: #F3F3F3;
-                        // margin-top: -4px;
-                        // margin-bottom: -4px;
-                    }
-
-                    /deep/ .q-stepper__tab {
-                        .q-stepper__title {
-                            font-size: 20px;
-                            font-family: $Titillium;
-                            font-weight: $bold;
-                            color: #2A2A2A;
-                        }
-
-                        &.q-stepper__tab--active,
-                        &.q-stepper__tab--done {
-                            .q-stepper__title {
-                                color: #7272FA;
-                            }
-
-                            .q-stepper__dot {
-                                background: #7272FA;
-                            }
-                        }
-                    }
-
-                    .--input {
-                        margin-top: 20px;
-
-                        /deep/ .q-field {
-                            height: 40px;
-                        }
-
-                        /deep/ .q-field__native,
-                        /deep/ .q-field__prefix,
-                        /deep/ .q-field__suffix {
-                            padding-top: 10px;
-                            padding-bottom: 0px;
-                        }
-
-                        /deep/ .q-field__label {
-                            top: 10px;
-                            font-size: 12px;
-                            font-weight: $bold;
-                            font-family: $Titillium;
-                        }
-
-                        /deep/ .q-field__marginal {
-                            height: 40px;
-                        }
-
-                        /deep/ .q-field__control {
-                            height: 40px;
-                        }
-                    }
-
-                    .--slider {
-                        /deep/ &.q-slider--dark {
-                            .q-slider__track-container {
-                                background: rgba(0, 0, 0, 0.3);
-                            }
-
-                            .q-slider__pin-value-marker-text {
-                                font-weight: $bold;
-                                font-size: 11px;
-                            }
-
-                            .q-slider__pin-value-marker-bg {
-                                background: #FFB200 !important;
-                            }
-
-                            .text-green {
-                                background: #FFB200 !important;
-                            }
-                        }
-                    }
-
-                    .--next-btn {
-                        width: 100px;
-                        text-transform: initial !important;
-                    }
-
-                    .--progress {
-                        height: 20px;
-
-                        /deep/ .q-linear-progress {
-                            margin-top: 8px;
-                            height: 5px !important;
-                            max-width: 90%;
-                            margin-left: auto;
-                            margin-right: auto;
-
-                            .q-linear-progress__track {
-                                background: #FFB200;
-                            }
-
-                            .q-linear-progress__model--indeterminate:before,
-                            .q-linear-progress__model--indeterminate:after {
-                                background: #FFB200;
-                            }
-                        }
-                    }
-
-                    .--back-btn {
-                        position: absolute;
-                        right: 0px;
-                        top: 6px;
-
-                        @media screen and (min-width: 1024px) {
-                            right: unset;
-                            left: 5px;
-                            top: 5px;
-                            width: 90px;
-
-                            /deep/ .q-focus-helper {
-                                display: none !important;
-                            }
-
-                            /deep/ .q-btn__wrapper {
-                                padding-left: 0px;
-                                padding-right: 0px;
-                            }
-                        }
-                    }
-
-                    .--subtitle {
-                        font-size: 16px;
-                        color: #000;
-                        font-family: $Titillium;
-                        font-weight: $light;
-                        line-height: 20px;
-                        margin-top: 10px;
-                        margin-bottom: 10px;
-
-                        ul {
-                            padding: 0px;
-                            margin: 0px;
-                            margin-left: 20px;
-
-                            li {
-                                font-size: 15px;
-                                font-weight: $light;
-                                margin-bottom: 10px;
-                                line-height: 21px;
-                                color: #FFB200;
-
-                                span {
-                                    color: #2A2A2A;
-                                }
-                            }
-                        }
-
-                        &__success {
-                            color: #00D0CA;
-                            font-weight: $bold;
-                            margin-bottom: 20px;
-                        }
-
-                        &__faild {
-                            color: #FFB200;
-                            font-weight: $bold;
-                            margin-bottom: 20px;
-                        }
-
-                        &__transLink {
-                            color: #2A2A2A;
-                            border-bottom: 1px solid;
-                            width: fit-content;
-                            font-weight: $bold;
-                            margin-bottom: 20px;
-                        }
-
-                        &__summary {
-                            margin-bottom: 20px;
-                            font-weight: $bold;
-                        }
-
-                        &__summary--list {
-                            list-style: disc;
-                            padding-left: 24px;
-                            margin-top: -10px;
-
-                            li {
-                                color: #B0B0B0;
-                            }
-                        }
-
-                        &__success {
-                            color: #00D0CA;
-                            font-weight: $bold;
-                            margin-bottom: 20px;
-                        }
-
-                        &__faild {
-                            color: #FFB200;
-                            font-weight: $bold;
-                            margin-bottom: 20px;
-                        }
-
-                        &__transLink {
-                            color: #2A2A2A;
-                            border-bottom: 1px solid;
-                            width: fit-content;
-                            font-weight: $bold;
-                            margin-bottom: 20px;
-                        }
-
-                        &__summary {
-                            margin-bottom: 20px;
-                            font-weight: $bold;
-                        }
-
-                        &__summary--list {
-                            list-style: disc;
-                            padding-left: 24px;
-                            margin-top: -10px;
-
-                            li {
-                                color: #B0B0B0;
-                            }
-                        }
-                    }
-
-                    .--title,
-                    .--amount {
-                        font-size: 13px;
-                        font-family: $Titillium;
-                        font-weight: $bold;
-                        color: #B0B0B0;
-                        text-transform: initial !important;
-                        line-height: 20px;
-                    }
-
-                    .--amount {
-                        color: #2A2A2A !important;
-                    }
-                }
-            }
-        }
-
-        &.open {
-            margin-bottom: 0px;
-            padding-left: 6%;
-            padding-right: 6%;
-
-            .list-wrapper {
-                visibility: visible;
-                height: auto;
-                opacity: 1;
-                transform: translateY(0px) scaleY(1);
-                margin-bottom: 10px;
-            }
-        }
-    }
-}
-
-.q-card {
-    border-radius: 25px;
-    box-shadow: 0 2px 4px -1px rgba(0, 0, 0, 0.2), 0 4px 35px rgba(0, 0, 0, 0.14), 0 1px 10px rgba(0, 0, 0, 0.12);
-}
-
-.accept-disclaimer {
-    color: #FFF !important;
-    background-color: #00D0DF !important;
-    text-transform: initial !important;
-    padding: 10px 30px;
-    border-radius: 50px;
-    font-weight: $light;
-}
-
-.parg {
-    font-size: 16px;
-    font-weight: $regular;
-    font-family: $Titillium;
-    line-height: 22px;
-
-    a {
-        font-weight: $bold;
-        border-bottom: 1px solid;
-        text-decoration: none;
-    }
-}
-
-.lower {
-    text-transform: initial !important;
-}
-
-.bold {
-    font-weight: $bold !important;
-}
-
-.stepper--desktop {
-    /deep/ .q-stepper__content {
-        max-width: 600px;
-        margin: auto;
-        margin-top: 50px;
-
-        .q-stepper__step-inner {
-            padding-top: 60px;
-        }
-    }
-
-    /deep/ .q-stepper__header {
-        // padding: 0px 10%;
-        background-color: #F6F6F6;
-        margin-bottom: 0px;
-        border-radius: 40px;
-        height: 40px;
-        position: relative;
-
-        &:before {
-            content: '';
-            width: 12%;
-            position: absolute;
-            left: 0px;
-            top: 18px;
-            height: 3px;
-            background-color: #F6F6F6;
-            transform: translateX(-100%);
-        }
-
-        &:after {
-            content: '';
-            width: 12%;
-            position: absolute;
-            right: 0px;
-            top: 18px;
-            height: 3px;
-            background-color: #F6F6F6;
-            transform: translateX(100%);
-        }
-
-        .q-stepper__tab {
-            height: 40px;
-            min-height: 40px;
-            padding: 0px 10px;
-            display: flex;
-            flex-direction: row;
-            align-items: center;
-
-            &--active,
-            &--done {
-                .q-stepper__label {
-                    .q-stepper__title {
-                        font-size: 12px;
-                        font-weight: bold;
-                    }
-                }
-            }
-        }
-
-        .q-stepper__dot {
-            &:after {
-                display: none;
-            }
-        }
-
-        .q-stepper__label {
-            margin-top: 0px;
-            padding-left: 6px;
-
-            .q-stepper__title {
-                font-size: 12px;
-                font-weight: normal;
-            }
-        }
-    }
-}
-</style><style lang="scss" scoped>
-@import "~@/assets/styles/variables.scss";
-
-.trade-component {
-    position: relative;
-    // font-family: $Franklin;
-    font-size: 18px;
-    font-weight: $bold;
-    border: 1px solid #CCC;
-    border-radius: 10px;
-    //max-width: 500px;
-    padding-bottom: 10px;
-
-    .prototype {
-        // position: absolute;
-        z-index: 0;
-        left: 0px;
-        top: 0px;
-
-        .--subtitle {
-            font-size: 14px;
-            // font-family: $Franklin;
-            font-weight: $light;
-            line-height: 20px;
-            margin-bottom: 20px;
-            text-align: center;
-            padding-left: 10px;
-            padding-right: 10px;
-        }
-
-        /deep/ .q-field--with-bottom {
-            .q-field__messages.col div {
-                padding-bottom: 0px;
-                margin-bottom: 0px;
-                margin-top: -8px;
-                font-size: 10px;
-                margin-left: -13px;
-            }
-        }
-
-        .head {
-            padding: 16px 30px;
-            border-bottom: 1px solid #f6f6f9;
-            margin-bottom: 15px;
-            font-size: 16px;
-        }
-
-        .you-pay {
-            &-head {
-                padding: 0px 30px;
-                margin-bottom: 18px;
-                font-size: 16px;
-
-                .col {
-                    &.red {
-                        font-size: 12px;
-                        font-weight: $light;
-                    }
-                }
-            }
-
-            &-body {
-                padding: 0px 30px;
-                font-size: 24px;
-                font-weight: $bold;
-            }
-        }
-
-        .choose-coin {
-            cursor: pointer;
-
-            &:hover {
-                opacity: .7;
-            }
-
-            position: relative;
-        }
-
-        .pay-coin {
-            position: relative;
-
-            .pay-coin-select {
-                position: absolute;
-                left: -29px;
-                top: 40px;
-                z-index: -10;
-                width: 348px;
-                opacity: 0;
-            }
-
-            .receive-coin-select {
-                position: absolute;
-                left: -29px;
-                top: 40px;
-                z-index: -10;
-                width: 348px;
-                opacity: 0;
-            }
-        }
-
-        .you-receive {
-            background-color: #f6f6f9;
-            margin: 0px 10px;
-            margin-top: 20px;
-            padding: 15px 10px;
-            border-radius: 10px;
-            position: relative;
-
-            /deep/ .chose_accounts {
-                background-color: #dfdff1 !important;
-                font-weight: $bold;
-                font-size: 14px !important;
-                // font-family: $Franklin !important;
-                color: #5e5e88 !important;
-                border-radius: 10px;
-                margin-top: 20px;
-                height: 54px;
-            }
-
-            /deep/.swap_vert {
-                position: absolute;
-                top: 0px;
-                left: 20px;
-                transform: translateY(-50%) scale(1.15);
-                background-color: #FFF !important;
-
-                .q-btn__wrapper:before {
-                    border: 3px solid #f4f4f4;
-                }
-            }
-
-            &-head {
-                padding: 0px 15px;
-                margin-bottom: 18px;
-                font-size: 16px;
-
-                .col {
-                    &.small {
-                        font-size: 11px;
-                        font-weight: $light;
-                    }
-
-                    &._loading {
-                        -webkit-animation: infiniteRotate 1s linear infinite;
-                        /* Safari */
-                        animation: infiniteRotate 1s linear infinite;
-                    }
-                }
-            }
-
-            &-body {
-                padding: 0px 15px;
-                font-size: 24px;
-                font-weight: $bold;
-            }
-        }
-    }
-}
-
-.info_rate_holder {
-    .i-btn {
-        &:hover {
-            .info_rate_wrapper {
-                visibility: visible;
-                opacity: 1;
-                transform: translateY(-10px);
-            }
-        }
-    }
-}
-
-.info_rate {
-    position: relative;
-
-    .i-btn {
-        width: 20px;
-        height: 20px;
-        border-radius: 20px;
-        border: 1px solid rgba(black, .2);
-        display: flex;
-        justify-content: center;
-        margin-left: 10px;
-    }
-
-    .rate-info-prototype {
-        position: absolute;
-        left: -220px;
-        top: 0px;
-        z-index: 9;
-        display: none;
-    }
-
-    .info_rate_wrapper {
-        position: absolute;
-        right: 0px;
-        top: 30px;
-        z-index: 9;
-        width: 220px;
-        background-color: #fff;
-        padding: 10px;
-        border-radius: 5px;
-        box-shadow: 0px 10px 10px 0px rgba(black, .1);
-        // font-family: $Franklin;
-        font-weight: $regular;
-        visibility: hidden;
-        opacity: 0;
-        transform: translateY(0px);
-        transition: ease opacity .3s, transform ease .3s;
-
-        hr {
-            opacity: .2;
-        }
-
-        .title {
-            font-size: 15px;
-            padding-top: 5px;
-            padding-bottom: 5px;
-            // font-weight: $semibold;
-        }
-    }
-}
-
-.disclaimer-wrapper {
-    font-size: 14px;
-    // font-family: $Franklin;
-    font-weight: $light;
-}
-
-/* Safari 4.0 - 8.0 */
-@-webkit-keyframes infiniteRotate {
-    0% {
-        opacity: 1;
-    }
-
-    100% {
-        opacity: .3;
-    }
-}
-
-/* Standard syntax */
-@keyframes infinite-rotate {
-    0% {
-        opacity: 1;
-    }
-
-    100% {
-        opacity: .3;
-    }
-}
-</style><style lang="scss">
-.option--avatar {
-    border: 1px solid rgba(black, .1);
-    width: 35px;
-    height: 35px;
-    max-width: 40px;
-    padding: 0px;
-    min-width: unset;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    border-radius: 50px;
-    overflow: hidden;
-
-    &__custom {
-        border: 1px solid;
-        margin-left: -10px;
-        padding-right: 0px;
-    }
-
-    &.vtx {
-        background-color: #000;
-        padding: 3px;
-    }
-}
-
-.q-field__messages {
-    margin-top: 5px;
-}
-</style>
